### 2018-07-26 3e4abca
tempdir check argument was only introduced in 3.5 R. So have made tempdir_check that checks R version before using argument.
