### 2018-07-26 #238 @OJWatson
Hi @annakrystalli - thank you for checking this through.

* Have just had a look and I didn't notice that ```tempdir(check)``` was only introduced in R 3.5, so I've just pushed some changes to catch for the R version before using this argument for backward compatibility.
* Have also trimmed those lines now.

Also a lot of the tests skip as you need the login details for the dummy account I've set up with the DHS website. These are stored within ```rdhs.json.tar.enc``` and are unlocked with a key that I've stored within my travis account. I can send the key if it would be helpful, or if there is a better way to share these securely then let me know as I'm still quite new to this sort of stuff.

Thank you again,

OJ
