### 2018-07-26 #238 @annakrystalli
Thanks for the quick response @OJWatson!

So I re-cloned the repo but I'm still getting errors during the tests. I include the output below, with some questions at the end of the comment:

#### check

``` r
pkg_dir <- "/Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs"
devtools::check(pkg_dir)

#> R CMD check results
#> 1 error  | 0 warnings | 1 note
#> checking tests ... ERROR
#>   Running ‘testthat.R’ [16s/486s]
#> Running the tests in ‘tests/testthat.R’ failed.
#> Last 13 lines of output:
#>   downloaded 80 KB
#>
#>   ══ testthat results  ════════════════════════════════════════════════════
#>   OK: 64 SKIPPED: 21 FAILED: 8
#>   1. Error: geojson works (@test_api_endpoints.R#60)
#>   2. Error: ETAR71FL.ZIP test (@test_downloads.R#139)
#>   3. Error: ugir41fl.zip test (@test_downloads.R#148)
#>   4. Error: zip file ending test (@test_downloads.R#157)
#>   5. Error: Hierarchal and sas7bdat dataset test (@test_downloads.R#168)
#>   6. Error: Large India Files (@test_downloads.R#214)
#>   7. Error: Odd nesting India (@test_downloads.R#223)
#>   8. Error: surveyId in extract (@test_extraction.R#109)
#>
#>   Error: testthat unit tests failed
#>   Execution halted
#>
#> checking R code for possible problems ... NOTE
#> tempdir_check: possible error in tempdir(TRUE): unused argument (TRUE)
```

Created on 2018-07-26 by the [reprex package](http://reprex.tidyverse.org) (v0.2.0).

#### test
``` r
pkg_dir <- "/Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs"
devtools::test(pkg_dir)
#> Loading rdhs
#> Loading required package: testthat
#> Testing rdhs
#> Warning: `encoding` is deprecated; all files now assumed to be UTF-8
#> ✔ | OK F W S | Context
#>
⠏ |  0       | API calls
⠋ |  0     1 | API calls
✔ |  0     1 | API calls
#> ─────────────────────────────────────────────────────────────────────────
#> test_api_calls.R:4: skip: can request api through dhs_api_request via client
#> No authentication available
#> ─────────────────────────────────────────────────────────────────────────
#>
⠏ |  0       | API endpoints
⠋ |  0     1 | API endpoints
⠙ |  0 1   1 | API endpoints
⠹ |  1 1   1 | API endpoints
⠸ |  2 1   1 | API endpoints
⠼ |  3 1   1 | API endpoints
⠴ |  4 1   1 | API endpoints
⠦ |  5 1   1 | API endpoints
⠧ |  6 1   1 | API endpoints
⠇ |  7 1   1 | API endpoints
⠏ |  8 1   1 | API endpoints
⠋ |  9 1   1 | API endpoints
⠙ | 10 1   1 | API endpoints
⠹ | 11 1   1 | API endpoints
⠸ | 12 1   1 | API endpoints
⠼ | 13 1   1 | API endpoints
⠴ | 14 1   1 | API endpoints
⠦ | 15 1   1 | API endpoints
⠧ | 16 1   1 | API endpoints
⠇ | 17 1   1 | API endpoints
⠏ | 18 1   1 | API endpoints
⠋ | 19 1   1 | API endpoints
⠙ | 20 1   1 | API endpoints
⠹ | 21 1   1 | API endpoints
⠸ | 22 1   1 | API endpoints
⠼ | 23 1   1 | API endpoints
⠴ | 24 1   1 | API endpoints
⠦ | 25 1   1 | API endpoints
⠧ | 26 1   1 | API endpoints
⠇ | 27 1   1 | API endpoints
⠏ | 28 1   1 | API endpoints
⠋ | 29 1   1 | API endpoints
✖ | 29 1   1 | API endpoints [303.2 s]
#> ─────────────────────────────────────────────────────────────────────────
#> test_api_endpoints.R:8: skip: format catches and al_results tests
#> No authentication available
#>
#> test_api_endpoints.R:60: error: geojson works
#> argument is of length zero
#> 1: dhs_data(countryIds = "SN", surveyYearStart = 2014, breakdown = "subnational",
#>        returnGeometry = TRUE, f = "geojson") at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/tests/testthat/test_api_endpoints.R:60
#> 2: handle_api_request(endpoint, query, all_results, client, force) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/api_endpoints.R:135
#> 3: api_request(endpoint, query, all_results) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/API.R:48
#> 4: handle_pagination_geojson(endpoint, query, all_results) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/API.R:95
#> ─────────────────────────────────────────────────────────────────────────
#>
⠏ |  0       | Authentication
⠋ |  0     1 | Authentication
⠙ |  0     2 | Authentication
✔ |  0     2 | Authentication
#> ─────────────────────────────────────────────────────────────────────────
#> test_authentication.R:5: skip: authenticate_dhs works
#> No authentication available
#>
#> test_authentication.R:24: skip: available_surveys works
#> No authentication available
#> ─────────────────────────────────────────────────────────────────────────
#>
⠏ |  0       | Client Setup
⠋ |  0     1 | Client Setup
⠙ |  0     2 | Client Setup
✔ |  0     2 | Client Setup
#> ─────────────────────────────────────────────────────────────────────────
#> test_client_set_up.R:5: skip: save credentials
#> No authentication available
#>
#> test_client_set_up.R:30: skip: new updates are recognised
#> No authentication available
#> ─────────────────────────────────────────────────────────────────────────
#>
⠏ |  0       | Downloads
⠋ |  0     1 | Downloads
⠙ |  0   1 1 | Downloads
⠹ |  0 1 1 1 | Downloads
⠸ |  0 1 2 1 | Downloads
⠼ |  0 2 2 1 | Downloads
⠴ |  0 2 3 1 | Downloads
⠦ |  0 3 3 1 | Downloads
⠧ |  0 3 4 1 | Downloads
⠇ |  0 4 4 1 | Downloads
⠏ |  0 4 4 2 | Downloads
⠋ |  0 4 5 2 | Downloads
⠙ |  0 5 5 2 | Downloads
⠹ |  0 5 6 2 | Downloads
⠸ |  0 6 6 2 | Downloads
✖ |  0 6 6 2 | Downloads
#> ─────────────────────────────────────────────────────────────────────────
#> test_downloads.R:4: skip: available surveys and download work
#> No authentication available
#>
#> test_downloads.R:139: warning: ETAR71FL.ZIP test
#> cannot open file 'rdhs.json': No such file or directory
#>
#> test_downloads.R:139: error: ETAR71FL.ZIP test
#> cannot open the connection
#> 1: new_rand_client() at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/tests/testthat/test_downloads.R:139
#> 2: rdhs::client_dhs(api_key = "ICLSPH-527168", config = read_rdhs_config_file("rdhs.json"),
#>        root = td) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/tests/testthat/helper_authentication.R:25
#> 3: read_rdhs_config_file("rdhs.json") at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/client.R:18
#> 4: jsonlite::fromJSON(readChar(config_path, fsize), FALSE) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/config.R:283
#> 5: readChar(config_path, fsize) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/config.R:283
#> 6: file(con, "rb")
#>
#> test_downloads.R:148: warning: ugir41fl.zip test
#> cannot open file 'rdhs.json': No such file or directory
#>
#> test_downloads.R:148: error: ugir41fl.zip test
#> cannot open the connection
#> 1: new_rand_client() at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/tests/testthat/test_downloads.R:148
#> 2: rdhs::client_dhs(api_key = "ICLSPH-527168", config = read_rdhs_config_file("rdhs.json"),
#>        root = td) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/tests/testthat/helper_authentication.R:25
#> 3: read_rdhs_config_file("rdhs.json") at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/client.R:18
#> 4: jsonlite::fromJSON(readChar(config_path, fsize), FALSE) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/config.R:283
#> 5: readChar(config_path, fsize) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/config.R:283
#> 6: file(con, "rb")
#>
#> test_downloads.R:157: warning: zip file ending test
#> cannot open file 'rdhs.json': No such file or directory
#>
#> test_downloads.R:157: error: zip file ending test
#> cannot open the connection
#> 1: new_rand_client() at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/tests/testthat/test_downloads.R:157
#> 2: rdhs::client_dhs(api_key = "ICLSPH-527168", config = read_rdhs_config_file("rdhs.json"),
#>        root = td) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/tests/testthat/helper_authentication.R:25
#> 3: read_rdhs_config_file("rdhs.json") at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/client.R:18
#> 4: jsonlite::fromJSON(readChar(config_path, fsize), FALSE) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/config.R:283
#> 5: readChar(config_path, fsize) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/config.R:283
#> 6: file(con, "rb")
#>
#> test_downloads.R:168: warning: Hierarchal and sas7bdat dataset test
#> cannot open file 'rdhs.json': No such file or directory
#>
#> test_downloads.R:168: error: Hierarchal and sas7bdat dataset test
#> cannot open the connection
#> 1: new_rand_client() at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/tests/testthat/test_downloads.R:168
#> 2: rdhs::client_dhs(api_key = "ICLSPH-527168", config = read_rdhs_config_file("rdhs.json"),
#>        root = td) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/tests/testthat/helper_authentication.R:25
#> 3: read_rdhs_config_file("rdhs.json") at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/client.R:18
#> 4: jsonlite::fromJSON(readChar(config_path, fsize), FALSE) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/config.R:283
#> 5: readChar(config_path, fsize) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/config.R:283
#> 6: file(con, "rb")
#>
#> test_downloads.R:189: skip: Geo dataset test
#> No authentication available
#>
#> test_downloads.R:214: warning: Large India Files
#> cannot open file 'rdhs.json': No such file or directory
#>
#> test_downloads.R:214: error: Large India Files
#> cannot open the connection
#> 1: new_rand_client() at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/tests/testthat/test_downloads.R:214
#> 2: rdhs::client_dhs(api_key = "ICLSPH-527168", config = read_rdhs_config_file("rdhs.json"),
#>        root = td) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/tests/testthat/helper_authentication.R:25
#> 3: read_rdhs_config_file("rdhs.json") at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/client.R:18
#> 4: jsonlite::fromJSON(readChar(config_path, fsize), FALSE) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/config.R:283
#> 5: readChar(config_path, fsize) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/config.R:283
#> 6: file(con, "rb")
#>
#> test_downloads.R:223: warning: Odd nesting India
#> cannot open file 'rdhs.json': No such file or directory
#>
#> test_downloads.R:223: error: Odd nesting India
#> cannot open the connection
#> 1: new_rand_client() at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/tests/testthat/test_downloads.R:223
#> 2: rdhs::client_dhs(api_key = "ICLSPH-527168", config = read_rdhs_config_file("rdhs.json"),
#>        root = td) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/tests/testthat/helper_authentication.R:25
#> 3: read_rdhs_config_file("rdhs.json") at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/client.R:18
#> 4: jsonlite::fromJSON(readChar(config_path, fsize), FALSE) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/config.R:283
#> 5: readChar(config_path, fsize) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/config.R:283
#> 6: file(con, "rb")
#> ─────────────────────────────────────────────────────────────────────────
#>
⠏ |  0       | Extraction
⠋ |  0     1 | Extraction
⠙ |  0   1 1 | Extraction
⠹ |  0 1 1 1 | Extraction
⠸ |  0 1 1 2 | Extraction
✖ |  0 1 1 2 | Extraction
#> ─────────────────────────────────────────────────────────────────────────
#> test_extraction.R:5: skip: query codes having downloaded surveys
#> No authentication available
#>
#> test_extraction.R:109: warning: surveyId in extract
#> cannot open file 'rdhs.json': No such file or directory
#>
#> test_extraction.R:109: error: surveyId in extract
#> cannot open the connection
#> 1: new_rand_client() at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/tests/testthat/test_extraction.R:109
#> 2: rdhs::client_dhs(api_key = "ICLSPH-527168", config = read_rdhs_config_file("rdhs.json"),
#>        root = td) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/tests/testthat/helper_authentication.R:25
#> 3: read_rdhs_config_file("rdhs.json") at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/client.R:18
#> 4: jsonlite::fromJSON(readChar(config_path, fsize), FALSE) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/config.R:283
#> 5: readChar(config_path, fsize) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/config.R:283
#> 6: file(con, "rb")
#>
#> test_extraction.R:125: skip: rbind_labelled
#> No authentication available
#> ─────────────────────────────────────────────────────────────────────────
#>
⠏ |  0       | Miscellaneous utils
⠋ |  1       | Miscellaneous utils
⠙ |  2       | Miscellaneous utils
⠹ |  3       | Miscellaneous utils
⠸ |  4       | Miscellaneous utils
⠼ |  5       | Miscellaneous utils
⠴ |  6       | Miscellaneous utils
⠦ |  7       | Miscellaneous utils
⠧ |  8       | Miscellaneous utils
⠇ |  8     1 | Miscellaneous utils
⠏ |  9     1 | Miscellaneous utils
⠋ |  9   1 1 | Miscellaneous utils
⠙ | 10   1 1 | Miscellaneous utils
✔ | 10   1 1 | Miscellaneous utils [0.1 s]
#> ─────────────────────────────────────────────────────────────────────────
#> test_miscellaneous.R:38: skip: slow api response
#> No authentication available
#>
#> test_miscellaneous.R:64: warning: different locales
#> OS reports request to set locale to "French_Belgium.1252" cannot be honored
#> ─────────────────────────────────────────────────────────────────────────
#>
⠏ |  0       | Parse datasets
⠋ |  1       | Parse datasets
⠙ |  2       | Parse datasets
⠹ |  3       | Parse datasets
⠸ |  4       | Parse datasets
⠼ |  5       | Parse datasets
⠴ |  6       | Parse datasets
⠦ |  7       | Parse datasets
⠧ |  8       | Parse datasets
⠇ |  9       | Parse datasets
⠏ | 10       | Parse datasets
⠋ | 11       | Parse datasets
⠙ | 12       | Parse datasets
⠹ | 13       | Parse datasets
⠸ | 14       | Parse datasets
⠼ | 15       | Parse datasets
⠴ | 16       | Parse datasets
⠦ | 17       | Parse datasets
⠧ | 17     1 | Parse datasets
✔ | 17     1 | Parse datasets [6.9 s]
#> ─────────────────────────────────────────────────────────────────────────
#> test_read_dhs_flat.R:75: skip: lower case flat file check
#> No authentication available
#> ─────────────────────────────────────────────────────────────────────────
#>
⠏ |  0       | user interface
⠋ |  0     1 | user interface
⠙ |  0     2 | user interface
⠹ |  0     3 | user interface
⠸ |  0     4 | user interface
⠼ |  0     5 | user interface
⠴ |  0     6 | user interface
⠦ |  0     7 | user interface
⠧ |  0     8 | user interface
⠇ |  0     9 | user interface
✔ |  0     9 | user interface
#> ─────────────────────────────────────────────────────────────────────────
#> test_ui.R:4: skip: check_for_client
#> No authentication available
#>
#> test_ui.R:37: skip: check on the use of the default client for API caching in extract
#> No authentication available
#>
#> test_ui.R:67: skip: get_datasets with no temp first
#> No authentication available
#>
#> test_ui.R:97: skip: get_model_datasets
#> No authentication available
#>
#> test_ui.R:107: skip: get_downloaded_datasets
#> No authentication available
#>
#> test_ui.R:118: skip: get_available_datasets
#> No authentication available
#>
#> test_ui.R:128: skip: search_and_extract_dhs
#> No authentication available
#>
#> test_ui.R:146: skip: get_variable_labels
#> No authentication available
#>
#> test_ui.R:168: skip: get_variable_labels direct via client
#> No authentication available
#> ─────────────────────────────────────────────────────────────────────────
#>
#> ══ Results ══════════════════════════════════════════════════════════════
#> Duration: 310.6 s
#>
#> OK:       56
#> Failed:   8
#> Warnings: 8
#> Skipped:  21
```

Created on 2018-07-26 by the [reprex package](http://reprex.tidyverse.org) (v0.2.0).

#### questions

- There still seems to be tests that are failing. Are these primarily still because of authentication issues? My understanding was that you are skipping any tests that require authentication? Should these not be skipping?
- Re: getting me authentication, it is never recommended to share your own private key with anyone. What would be better is that I get an authorisation token for myself. I imagine any user of the package would need this too? In which case, the README should flag this and also include instructions on how to get an authorisation token and use it to access the API. You could also include a whole separate vignette for this but it should still be flagged and linked to prominently in the README
