### 2018-07-26 dd28916
adding auth skips to tests where ou need login credentials
