### 2018-07-26 a05bc3e
at the moment we just skip the geojson test always as their API is still too slow for these large data objects to be usful within test suites
