+++ tests/testthat/helper_authentication.R
@@ -7,12 +7,32 @@ skip_if_no_auth <- function(config_path = "rdhs.json") {
     skip("No authentication available")
   }
 
+  skip_if_slow_API()
+
+}
+
+skip_if_slow_API <- function() {
+
   if (suppressMessages(last_api_update()) == 0.5) {
     skip("API is down")
   }
 
 }
 
+
+api_timeout_safe_test <- function(expr) {
+
+  # one that has to paginate
+  d <- tryCatch(eval(substitute(expr)), error = function(e) NULL )
+
+  if (is.null(d)) {
+    testthat::skip(paste0("Skipping test as API is too ",
+                          "slow to return these objects."))
+  }
+
+  return(d)
+
+}
 
 
 new_rand_client <- function() {
