+++ README.Rmd
@@ -47,6 +47,8 @@ library(rdhs)
 
 * An example workflow using `rdhs` to calculate trends in anemia prevalence is available [here](https://ojwatson.github.io/rdhs/articles/anemia.html).
 
+* To run the full test suite (including tests for downloading datasets) see tutorial [here](https://ojwatson.github.io/rdhs/articles/testing.html).
+
 ## Basic Functionality
 
 ### Query the [DHS API](https://api.dhsprogram.com/).
