### 2018-07-27 #238 @OJWatson
Hi,

* I have just added another vignette on this [here](https://ojwatson.github.io/rdhs/articles/testing.html), which is linked to in the readme.

* I've also added some bubble-wrapping around the API tests, so that if they haven't responded after 30 seconds (default timeout) then these tests are skipped. And then not in tests, these will display more helpful error messages about why.

And yep happy for reviewers.

Thank you again for all the help,

OJ
