### 2018-08-01 0b313cd
change to checking dataset file names, so that we can pass in either FILENAME.Zip or FILENAME
