### 2018-08-01 1486c9a
better cache checking
