### 2018-08-01 2416b4d
cache last api update fix needs to query all the dhs_datasets endpoint without relyiong on client functionality (which starts by checking the updates and so is circular). Therefore query this manually
