### 2018-08-01 feeb0af
Crucial patch, to the client search functionality. As a result of changing the output from get_datasets to show when you don't get datasets back due to credentials, wekept these in the return object Thus if you feed that straight back it was introducing nas into the search steps causing errors when looping through.
