### 2018-08-13 b864916
change to the print config so that password is obscured
