### 2018-08-23 #238 @czeildi
Hi,

@OJWatson I have one small remark: the NEWS could be more concise, more shorter bullet points grouped according to new function / bug fix etc would make it easier to skim through it. You  may consider getting ideas from http://style.tidyverse.org/news.html.

FYI I won't be reviewing the package as discussed with Anna.
