### 2018-08-24 #238 @OJWatson
Hi,

Thanks for this @czeildi - have never written a NEWS section before, and reading it back I have definitely waffled. I'll trim and tidy it up in line with the style guide you've linked.

Many thanks, OJ
