+++ vignettes/testing.Rmd
@@ -16,7 +16,7 @@ The following is a guide describing how to be able to run the full test suite lo
 
 ## 1. Creating an account with the DHS website
 
-In order to run all the tests in the *rdhs* package you need to have created an account with the DHS website. This requires about 5 mins to fill in their form and then a few days for them to approve your account creation. To start this head to the DHS website and go to the Data tab [here](https://dhsprogram.com/data/dataset_admin/login_main.cfm). A full rationale for this procss can be found on their website [her](https://dhsprogram.com/data/Access-Instructions.cfm). In short you provide your contact information and then describe the stuy you are conducting that requires the DHS survey datasets.
+In order to run all the tests in the *rdhs* package you need to have created an account with the DHS website. This requires about 5 mins to fill in their form and then a few days for them to approve your account creation. To start this head to the DHS website and go to the Data tab [here](https://dhsprogram.com/data/dataset_admin/login_main.cfm). A full rationale for this procss can be found on their website [here](https://dhsprogram.com/data/Access-Instructions.cfm). In short you provide your contact information and then describe the study you are conducting that requires the DHS survey datasets.
 
 Importantly, when filling in their form describing the description of your study be sure to include some mention of needing geographic datasets as a number of the tests download geographic datasets. After you have filled this in you need to specify which countries and datasets you need. In order for the full test suite to work you must request every country and dataset type possible. After you have done this the DHS programme will take a few days to approve you request and will then email you to let you know at the email address you have provided. 
 
