+++ man-roxygen/client_dhs.yml
@@ -49,7 +49,7 @@ get_datasets:
   short: >-
     Gets datasets from your cache or downloads from the DHS website. By providing the filenames, as specified in one of the
     returned fields from \code{\link{dhs_datasets}}, the client will log in for you and download all
-    the files you hae requested. If any of the requested files are unavailable for your log in, these
+    the files you have requested. If any of the requested files are unavailable for your log in, these
     will be flagged up first as a message so you can make a note and request them through the DHS website.
     You also have the option to control whether the downloaded zip file is then extracted and converted
     into a more convenient R \code{data.frame}. This converted object will then be subsequently saved as
@@ -135,7 +135,7 @@ extract:
     questions: >-
       Questions to be queried, in the format from \code{survey_questions}
     add_geo: >-
-      Add geograpic information to the extract. Defaut = TRUE
+      Add geographic information to the extract. Default = TRUE
 
 get_variable_labels:
   short: >-
