### 2018-09-20 #238 @OJWatson
Hi there @LucyMcGowan,

Thank you so much for the review, especially with your 👶 due so soon. Hope everything goes well.

I will respond fully by the end of today, with my changes but I agree with everything that you have suggested, in particular making sure everything has an example and then shortening the examples that the API functions run to just one function call.

Thank you again,

OJ
