### 2018-09-21 d66c8e5
Test suite assumed you had the encrypted rdhs.json file locally. new_rand_client() been changed to make a dummy one if this is not the case. this will enable the api tests to correctly run, and addresses reviewer comment (RC) 2.
