+++ tests/testthat/test_api_endpoints.R
@@ -13,18 +13,18 @@ test_that("format catches and all_results tests", {
   dat <- api_timeout_safe_test(
     dhs_data_updates(
       client = cli, lastUpdate = "20150901", all_results = FALSE
-    )
+    ), cli
   )
   dat2 <- api_timeout_safe_test(
     dhs_data_updates(
       client = cli, lastUpdate = "20150901", all_results = FALSE
-    )
+    ), cli
   )
   expect_identical(dat, dat2)
 
   # tets for non "json" specification
   dat <- api_timeout_safe_test(
-    dhs_datasets(f = "xml", all_results = FALSE)
+    dhs_datasets(f = "xml", all_results = FALSE), cli
   )
 
   # tests for rubbish arguments
@@ -33,28 +33,26 @@ test_that("format catches and all_results tests", {
       indicatorIds = "ML_FEVT_C_AMasfafasfL",
       surveyYearStart = 202231231306,
       breakdown = "subParTyping"
-    )
+    ), cli
   )
   )
   dat <- api_timeout_safe_test(
     dhs_datasets(
       surveyIds = "laksjdoash,dasjd", f = "xml", all_results = FALSE
-    )
+    ), cli
   )
 
   # test for misstyped args
-  expect_error(api_timeout_safe_test(
-    dhs_countries(countryIds = "SE"))
-  )
+  expect_error(api_timeout_safe_test(dhs_countries(countryIds = "SE"), cli))
 
   # test for all_results  and smaller than 5000 page
   dat <- api_timeout_safe_test(
-    dhs_indicators()
+    dhs_indicators(), cli
   )
 
   # test for all_results  and larger than 5000 page
   datasets <- api_timeout_safe_test(
-    dhs_datasets()
+    dhs_datasets(), cli
   )
 
   # test for data.table extension
@@ -63,7 +61,7 @@ test_that("format catches and all_results tests", {
   dat <- api_timeout_safe_test(
     dhs_countries(
       countryIds = "TZ", surveyYearStart = "2010", all_results = FALSE
-    )
+    ), cli
   )
   expect_true(inherits(dat, "data.table"))
 
@@ -82,7 +80,7 @@ test_that("geojson works", {
              surveyYearStart = 2014,
              breakdown = "subnational",
              returnGeometry = TRUE,
-             f = "geojson")
+             f = "geojson"), cli
   )
 
   expect_true(inherits(d, "list"))
@@ -96,7 +94,7 @@ test_that("geojson works", {
              surveyYearStart = 2014,
              breakdown = "subnational",
              returnGeometry = TRUE,
-             f = "geojson")
+             f = "geojson"), cli
   )
 
   expect_true(inherits(d, "list"))
@@ -114,7 +112,7 @@ test_that("dhs_countries works", {
   dat <- api_timeout_safe_test(
     dhs_countries(
       countryIds = "SN", surveyYearStart = "2010", all_results = FALSE
-    )
+    ), cli
   )
   expect_identical(dat$UNAIDS_CountryCode[1], "SEN")
   dat <- api_timeout_safe_test(
@@ -122,12 +120,12 @@ test_that("dhs_countries works", {
       countryIds = c("EG", "SN"),
       surveyYearStart = "1991", surveyYearEnd = "2006",
       surveyType = "DHS", surveyCharacteristicIds = "32"
-    )
+    ), cli
   )
   expect_identical(dat$ISO3_CountryCode[1:2], c("EGY", "SEN"))
[…]
