### 2018-09-21 3e8fc05
add add_line function from httr
