### 2018-09-21 1ba973e
config path now suitbaly added to gitignore and buildignore within write_rdhs_config_path. This addresses RC 4
