### 2018-09-21 6c55549
Added more information to the readme and to the introduction about how to create an account with the DHS website, and provided a link to the website's guide on how to create an account. This description has been placed at the beginning of the README and is then referenced in section 3 of the README. It has also been put in the relevant area in the introduction.Rmd. This commit addresses RC 5
