+++ README.md
@@ -11,6 +11,8 @@ Motivation
 
 The Demographic and Health Surveys (DHS) Program has collected population survey data from over 90 countries for over 30 years. In many countries, DHS provide the key data that mark progress towards targets such as the Sustainable Development Goals (SDGs) and inform health policy. Though standard health indicators are routinely published in survey final reports, much of the value of DHS is derived from the ability to download and analyse standardized microdata datasets for subgroup analysis, pooled multi-country analysis, and extended research studies. The suite of tools within `rdhs` improves the accessibility of these datasets for statistical analysis with R, with aim to support reproducible global health research and simplify common analytical pipelines.
 
+To be able to download survey datasets from the DHS website, you will need to set up an account with them to enable you to request access to the datasets. Instructions on how to do this can be found [here](https://dhsprogram.com/data/Access-Instructions.cfm). The email, password, and project name that were used to create the account will then need to be provided to `rdhs` when attempting to download datasets.
+
 ------------------------------------------------------------------------
 
 `rdhs` is a package for management and analysis of [Demographic and Health Survey (DHS)](www.dhsprogram.com) data. This includes functionality to:
@@ -46,7 +48,7 @@ Basic Functionality
 
 ### Query the [DHS API](https://api.dhsprogram.com/).
 
-Obtain survey esimates for Malaria prevalence among children from the Democratic Republic of Congo and Tanzania in the last 5 years (since 2013) that included rapid diagnostic tests (RDTs).
+Obtain survey estimates for Malaria prevalence among children from the Democratic Republic of Congo and Tanzania in the last 5 years (since 2013) that included rapid diagnostic tests (RDTs).
 
 ``` r
 dhs_indicators(indicatorIds = "ML_PMAL_C_RDT", returnFields=c("IndicatorId", "ShortName"))
@@ -96,23 +98,27 @@ datasets <- dhs_datasets(surveyIds = survs$SurveyId, fileFormat = "FL", fileType
 str(datasets)
 #> 'data.frame':    2 obs. of  13 variables:
 #>  $ FileFormat          : chr  "Flat ASCII data (.dat)" "Flat ASCII data (.dat)"
-#>  $ FileSize            : int  6595349 6622108
+#>  $ FileSize            : int  6595349 6622102
 #>  $ DatasetType         : chr  "Survey Datasets" "Survey Datasets"
 #>  $ SurveyNum           : int  421 485
 #>  $ SurveyId            : chr  "CD2013DHS" "TZ2015DHS"
 #>  $ FileType            : chr  "Household Member Recode" "Household Member Recode"
-#>  $ FileDateLastModified: chr  "September, 19 2016 09:58:23" "December, 09 2016 14:10:59"
+#>  $ FileDateLastModified: chr  "September, 19 2016 09:58:23" "August, 07 2018 17:36:25"
 #>  $ SurveyYearLabel     : chr  "2013-14" "2015-16"
 #>  $ SurveyType          : chr  "DHS" "DHS"
 #>  $ SurveyYear          : int  2013 2015
 #>  $ DHS_CountryCode     : chr  "CD" "TZ"
-#>  $ FileName            : chr  "CDPR61FL.ZIP" "TZPR7HFL.ZIP"
+#>  $ FileName            : chr  "CDPR61FL.ZIP" "TZPR7AFL.ZIP"
 #>  $ CountryName         : chr  "Congo Democratic Republic" "Tanzania"
 ```
 
 ### Download datasets
 
[…]
