+++ R/client.R
@@ -10,6 +10,25 @@
 #' @template client_dhs_methods
 #' @export
 #'
+#' @example
+#'
+#' #Create new directory
+#' td <- file.path(tempdir(), as.integer(Sys.time()))
+#'
+#' # create an rdhs config file at "rdhs.json
+#' set_rdhs_config(
+#' config_path = "rdhs.json", global = FALSE, password_prompt = FALSE
+#' )
+#'
+#' # create your client, specifying your config
+#' cli <- client_dhs(
+#' api_key = "ICLSPH-527168",
+#' config = read_rdhs_config_file("rdhs.json"),
+#' root = tempdir()
+#' )
+
+
+return(cli)
 client_dhs <- function(config=NULL,
                        root=rappdirs_rdhs(),
                        api_key="ICLSPH-527168") {
