### 2018-09-21 e3b6997
example for get_datasets and extract_dhs
