+++ R/read_datasets.R
@@ -243,6 +243,18 @@ get_labels_from_dataset <- function(data, return_all=TRUE) {
 #'   either as the file path or after having
 #'   been read using \code{readRDS}
 #' @export
+#'
+#' @example
+#' # get the model datasets included with the package
+#' model_datasets <- model_datasets
+#'
+#' # download one of them
+#' g <- get_datasets(dataset_filenames = model_datasets$FileName[1])
+#' dl <- data_and_labels(g$zzbr62dt)
+#'
+#' # now we easily have our survey question labels easily accessible
+#' grep("bed net", dl$variable_names$description, value = TRUE)
+#'
 data_and_labels <- function(dataset) {
 
   if (class(dataset)[1] == "character") {
