### 2018-09-21 eebb0ca
more examples to reading dataset functions
