+++ R/config.R
@@ -126,6 +126,21 @@ handle_config <- function(config_path) {
 #' @return Invisibly returns the rdhs config object
 #' @export
 #'
+#' @example
+#'
+#' \dontrun{
+#' # normal set up we would prvide the email and project, and be prompted for
+#' # the password. (not run as it requires a prompt)
+#' set_rdhs_config(email = "blah@gmail.com", project = "Blahs",
+#' config_path = "rdhs.json", global = FALSE)
+#'
+#' # you can look at what you have set these to using \code{get_rdhs_config}
+#' get_rdhs_config()
+#'
+#' # and you can view it prettily using \code{print_rdhs_config}
+#' print_rdhs_config(get_rdhs_config())
+#' }
+#'
 set_rdhs_config <- function(email = NULL,
                             project = NULL,
                             cache_path = NULL,
