+++ man/extract_dhs.Rd
@@ -10,7 +10,7 @@ extract_dhs(questions, add_geo = FALSE)
 \item{questions}{Questions to be queried, in the format from
 \code{\link{search_variables}} or \code{\link{search_variable_labels}}}
 
-\item{add_geo}{Add geograpic information to the extract. Defaut = `TRUE`}
+\item{add_geo}{Add geographic information to the extract. Defaut = `TRUE`}
 }
 \value{
 A \code{list} of `data.frames` for each survey data extracted.
@@ -23,4 +23,19 @@ requested survey variables or survey definitions
 Function to extract datasets using a set of survey questions as
   taken from the output from \code{\link{search_variables}}
   or \code{\link{search_variable_labels}}
+}
+\examples{
+
+# get the model datasets included with the package
+model_datasets <- model_datasets
+
+# download one of them
+g <- get_datasets(dataset_filenames = model_datasets$FileName[1])
+dl <- data_and_labels(g$zzbr62dt)
+
+# create some terms of data me may want to extrac
+st <- search_variable_labels(names(g), "bed net")
+
+# and now extract it
+ex <- extract_dhs(st)
 }
