### 2018-09-24 027b576
other exported function examples
