+++ man/read_zipdata.Rd
@@ -21,3 +21,26 @@ identified file within the zip. Default = `foreign::read.dta`}
 \description{
 Read filetype from a zipped folder based on the file ending
 }
+\examples{
+
+# get the model datasets included in the package
+model_datasets <- model_datasets
+
+# download just the zip
+g <- get_datasets(
+dataset_filenames = model_datasets$FileName[1],
+download_option = "zip"
+)
+
+# and then read from the zip. This function is used internally by rdhs
+# when using `get_datasets` with `download_option = .rds` (default)
+r <- read_zipdata(
+g[[1]], pattern = ".dta"
+)
+
+# and we can pass a function to read the file and any other args with ...
+r <- read_zipdata(
+g[[1]], pattern = ".dta", readfn = haven::read_dta, encoding = "UTF-8"
+)
+
+}
