+++ man/data_and_labels.Rd
@@ -15,3 +15,15 @@ been read using \code{readRDS}}
 Function to give the former output of get_datasets as it can be nice to have
 both the definitions and the dataset attached together
 }
+\examples{
+# get the model datasets included with the package
+model_datasets <- model_datasets
+
+# download one of them
+g <- get_datasets(dataset_filenames = model_datasets$FileName[1])
+dl <- data_and_labels(g$zzbr62dt)
+
+# now we easily have our survey question labels easily accessible
+grep("bed net", dl$variable_names$description, value = TRUE)
+
+}
