### 2018-09-24 41beb47
Better examples for API
