+++ R/api_endpoints.R
@@ -1,7 +1,7 @@
 #' API request of DHS Indicator Data
 #'
 #' @title API request of DHS Indicator Data
-#' @param client If the api request should be cached, then provide a client
+#' @param client If the API request should be cached, then provide a client
 #'   object created by \code{\link{client_dhs}}
 #' @param force Should we force fetching the API results, and ignore any
 #'   cached results we have. Default = FALSE
@@ -78,7 +78,23 @@
 #'
 #' @export
 #' @examples
-#' ## not run only because they take a very long time and interact with an API
+#'
+#' # A common use for the indicator data API will be to search for a specific
+#' # health indicator for a given country. For example to return the total
+#' # malaria prevalence according to RDT, given by the indicator ML_PMAL_C_RDT,
+#' # in Senegal since 2010:
+#'
+#' dat <- dhs_data(
+#' indicatorIds="ML_PMAL_C_RDT",
+#' countryIds="SN",
+#' surveyYearStart="2006"
+#' )
+#'
+#' # A complete list of examples for how each argument to the data api
+#' # endpoint can be provided is given below, which is a copy of each of
+#' # the examples listed in the API at:
+#' \url{https://api.dhsprogram.com/#/api-data.cfm}
+#'
 #' \dontrun{
 #' dat <- dhs_data(countryIds="EG",all_results=FALSE)
 #' dat <- dhs_data(indicatorIds="FE_FRTR_W_TFR",all_results=FALSE)
@@ -140,7 +156,7 @@ dhs_data <- function(countryIds=NULL,
 #' API request of DHS Indicators
 #'
 #' @title API request of DHS Indicators
-#' @param client If the api request should be cached, then provide a client
+#' @param client If the API request should be cached, then provide a client
 #'   object created by \code{\link{client_dhs}}
 #' @param force Should we force fetching the API results, and ignore any
 #'   cached results we have. Default = FALSE
@@ -187,8 +203,24 @@ dhs_data <- function(countryIds=NULL,
 #'   \url{https://api.dhsprogram.com/rest/dhs/indicators/fields}
 #'
 #' @export
+#'
 #' @examples
-#' ## not run only because they take a very long time and interact with an API
+#'
+#' # A common use for the indicators data API will be to search for a list of
+#' # health indicators within a given characteristic category, such as anemia
+#' # testing, HIV prevalence, micronutrients etc. For example to return all the
+#' # indicators relating to malaria testing by RDTs:
+#'
+#' dat <- dhs_indicators(surveyCharacteristicIds="90")
+#'
+#' # A list of the different `surveyCharacteristicIds` can be found
+#' # [here](https://api.dhsprogram.com/rest/dhs/surveycharacteristics?f=html)
+#'
+#' # A complete list of examples for how each argument to the indicator api
+#' # endpoint can be provided is given below, which is a copy of each of
+#' # the examples listed in the API at:
+#' \url{https://api.dhsprogram.com/#/api-indicators.cfm}
+#'
 #' \dontrun{
 #' dat <- dhs_indicators(countryIds="EG",all_results=FALSE)
 #' dat <- dhs_indicators(indicatorIds="FE_FRTR_W_TFR",all_results=FALSE)
@@ -235,7 +267,7 @@ dhs_indicators <- function(countryIds=NULL,
 #' API request of DHS UI Updates
 #'
 #' @title API request of DHS UI Updates
-#' @param client If the api request should be cached, then provide a client
+#' @param client If the API request should be cached, then provide a client
 #'   object created by \code{\link{client_dhs}}
 #' @param force Should we force fetching the API results, and ignore any
 #'   cached results we have. Default = FALSE
@@ -260,7 +292,18 @@ dhs_indicators <- function(countryIds=NULL,
 #'
[…]
