### 2018-09-24 f067fae
ability to create a config in a directory that is not the current directory
