+++ R/config.R
@@ -110,7 +110,10 @@ handle_config <- function(config_path) {
 #'   `FALSE` the config file will be saved within the current directory. This
 #'   can be useful if you create a new DHS project for each new piece of work,
 #'   and want to keep the datasets you download for this project separate to
-#'   another.
+#'   another. If you want to have your config file saved in a different
+#'   directory, then you must create a file "rdhs.json" first in that directory
+#'   before specifying the full path to it, as well as setting `global` equal to
+#'   `FALSE`.
 #'
 #'   As an aside, it is useful for the DHS program to see how the surveys they
 #'   conducted are being used, and thus it is helpful for them if you do create
@@ -186,8 +189,10 @@ set_rdhs_config <- function(email = NULL,
         stop("For a global configuration, 'config_path' must be '~/.rdhs.json'")
       }
     } else {
-      if (config_path != "rdhs.json"){
-        stop("For a local configuration, 'config_path' must be 'rdhs.json'")
+      if (config_path != "rdhs.json" &&
+          (!file.exists(config_path) || basename(config_path) == "rdhs/json")) {
+        stop("For a local configuration, 'config_path' must be 'rdhs.json' ",
+             "or 'config_path' must already exist and end 'rdhs.json'")
       }
     }
   } else {
