+++ man/get_datasets.Rd
@@ -4,8 +4,9 @@
 \alias{get_datasets}
 \title{Get Datasets}
 \usage{
-get_datasets(dataset_filenames, download_option = "rds", reformat = FALSE,
-  all_lower = TRUE, output_dir_root = NULL, clear_cache = FALSE, ...)
+get_datasets(dataset_filenames, download_option = "rds",
+  reformat = FALSE, all_lower = TRUE, output_dir_root = NULL,
+  clear_cache = FALSE, ...)
 }
 \arguments{
 \item{dataset_filenames}{The desired filenames to be downloaded. These can be
