### 2018-09-24 570c9c3
pkgdown changes for new vignette
