+++ vignettes/client.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "rdhs client object and internal package design"
 author: "OJ Watson, Jeff Eaton"
-date: "2018-07-19"
+date: "2018-09-24"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Using the rdhs client}
@@ -123,8 +123,8 @@ data[1,]
 ```
 
 ```
-##    DataId                           Indicator  SurveyId IsPreferred Value
-## 1 2582633 Malaria prevalence according to RDT CD2013DHS           1  17.1
+##   DataId                           Indicator  SurveyId IsPreferred Value
+## 1 941297 Malaria prevalence according to RDT CD2013DHS           1  17.1
 ##        SDRID Precision        RegionId SurveyYearLabel SurveyType
 ## 1 MLPMALCRDT         1 CDDHS2013503010         2013-14        DHS
 ##   SurveyYear IndicatorOrder DHS_CountryCode CILow
@@ -163,7 +163,7 @@ ggplot(resp[resp$CountryName %in% countries,],
   facet_wrap(~CountryName,ncol = 6) 
 ```
 
-<img src="https://github.com/OJWatson/rdhs/raw/patching/vignettes/client_files/figure-html/unnamed-chunk-1-1.png" style="display: block; margin: auto;" />
+<img src="C:/Users/Oliver/GoogleDrive/AcademicWork/Imperial/git/rdhs/vignettes/client_files/figure-html/unnamed-chunk-1-1.png" style="display: block; margin: auto;" />
 
 
 If we incorrectly entered a filter query (very possible), `rdhs` will let us know our request was invalid:
@@ -235,7 +235,7 @@ str(datasets)
 ```
 ## 'data.frame':	19 obs. of  13 variables:
 ##  $ FileFormat          : chr  "Flat ASCII data (.dat)" "Flat ASCII data (.dat)" "Flat ASCII data (.dat)" "Flat ASCII data (.dat)" ...
-##  $ FileSize            : int  198561 7030083 3226262 8028957 11426382 4794941 1569680 6595349 63022 394824 ...
+##  $ FileSize            : int  198561 7030083 3226262 8028957 11426382 4794941 1569680 6595349 63022 996906 ...
 ##  $ DatasetType         : chr  "HIV Datasets" "Survey Datasets" "Survey Datasets" "Survey Datasets" ...
 ##  $ SurveyNum           : int  421 421 421 421 421 421 421 421 421 421 ...
 ##  $ SurveyId            : chr  "CD2013DHS" "CD2013DHS" "CD2013DHS" "CD2013DHS" ...
@@ -293,7 +293,7 @@ client
 ##   Private:
 ##     api_endpoints: data indicators countries surveys surveycharacteristics  ...
 ##     api_key: ICLSPH-527168
-##     cache_date: 2018-07-19 11:44:39
+##     cache_date: 2018-09-24 16:36:45
 ##     check_available_datasets: function (filenames) 
 ##     config: rdhs_config
 ##     na_s: ^na -|^na-|.*-na$|.* - na$| \{NA\}$|.* NA$|.*NA$
@@ -378,8 +378,8 @@ table(questions$dataset_filename)
 
 ```
 ## 
-## CDHR61FL CDPR61FL TZGC7AFL TZIR7HFL 
-##       24        1        1       48
+## CDHR61FL CDPR61FL TZHR7AFL TZPR7AFL 
[…]
