+++ vignettes/introduction.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "How to use rdhs?"
 author: "OJ Watson, Jeff Eaton"
-date: "2018-07-19"
+date: "2018-09-24"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{How to use rdhs?}
@@ -115,8 +115,8 @@ data[1,]
 ```
 
 ```
-##    DataId                           Indicator  SurveyId IsPreferred Value
-## 1 2582633 Malaria prevalence according to RDT CD2013DHS           1  17.1
+##   DataId                           Indicator  SurveyId IsPreferred Value
+## 1 941297 Malaria prevalence according to RDT CD2013DHS           1  17.1
 ##        SDRID Precision        RegionId SurveyYearLabel SurveyType
 ## 1 MLPMALCRDT         1 CDDHS2013503010         2013-14        DHS
 ##   SurveyYear IndicatorOrder DHS_CountryCode CILow
@@ -155,7 +155,7 @@ ggplot(resp[resp$CountryName %in% countries,],
   facet_wrap(~CountryName,ncol = 6) 
 ```
 
-<img src="https://github.com/OJWatson/rdhs/raw/patching/vignettes/introduction_files/figure-html/unnamed-chunk-1-1.png" style="display: block; margin: auto;" />
+<img src="C:/Users/Oliver/GoogleDrive/AcademicWork/Imperial/git/rdhs/vignettes/introduction_files/figure-html/unnamed-chunk-1-1.png" style="display: block; margin: auto;" />
 
 
 If we incorrectly entered a filter query (very possible), `rdhs` will let us know our request was invalid:
@@ -230,7 +230,7 @@ str(datasets)
 ```
 ## 'data.frame':	19 obs. of  13 variables:
 ##  $ FileFormat          : chr  "Flat ASCII data (.dat)" "Flat ASCII data (.dat)" "Flat ASCII data (.dat)" "Flat ASCII data (.dat)" ...
-##  $ FileSize            : int  198561 7030083 3226262 8028957 11426382 4794941 1569680 6595349 63022 394824 ...
+##  $ FileSize            : int  198561 7030083 3226262 8028957 11426382 4794941 1569680 6595349 63022 996906 ...
 ##  $ DatasetType         : chr  "HIV Datasets" "Survey Datasets" "Survey Datasets" "Survey Datasets" ...
 ##  $ SurveyNum           : int  421 421 421 421 421 421 421 421 421 421 ...
 ##  $ SurveyId            : chr  "CD2013DHS" "CD2013DHS" "CD2013DHS" "CD2013DHS" ...
@@ -250,7 +250,9 @@ We can now use this to download our datasets for further analysis.
 
 ## 3. Download survey datasets
 
-We can now go ahead and download our datasets. To do this we need to first set our login credentials for the DHS website. Once this is set you will be able to download the datasets that you have access to. In order to set up our credentials we use the function `set_rdhs_config()`. This will require providing as arguments your `email` and `project` for which you want to download datasets from. You will then be prompted for your password. 
+We can now go ahead and download our datasets. To be able to download survey datasets from the DHS website, you will need to set up an account with them to enable you to request access to the datasets. Instructions on how to do this can be found [here](https://dhsprogram.com/data/Access-Instructions.cfm). The email, password, and project name that were used to create the account will then need to be provided to `rdhs` when attempting to download datasets. 
+
[…]
