### 2018-09-24 7c87f8c
pkgdown corrections
