+++ R/api_endpoints.R
@@ -71,7 +71,7 @@
 #' @param page Allows specifying a page number to obtain for the API request.
 #'   By default the API will return page 1.
 #'
-#' @return Returns a `data.table` of 27 (or less if `returnFields` is provided)
+#' @return Returns a \code{data.table} of 27 (or less if \code{returnFields} is provided)
 #'   data for your particular query. Details of properties returned with each
 #'   row of data are provided at
 #' \url{https://api.dhsprogram.com/rest/dhs/data/fields}
@@ -93,7 +93,8 @@
 #' # A complete list of examples for how each argument to the data api
 #' # endpoint can be provided is given below, which is a copy of each of
 #' # the examples listed in the API at:
-#' \url{https://api.dhsprogram.com/#/api-data.cfm}
+#'
+#' https://api.dhsprogram.com/#/api-data.cfm
 #'
 #' \dontrun{
 #' dat <- dhs_data(countryIds="EG",all_results=FALSE)
@@ -197,7 +198,7 @@ dhs_data <- function(countryIds=NULL,
 #' @param page Allows specifying a page number to obtain for the API request. By
 #'   default the API will return page 1.
 #'
-#' @return Returns a `data.table` of 18 (or less if `returnFields` is provided)
+#' @return Returns a \code{data.table} of 18 (or less if \code{returnFields} is provided)
 #'   indicators with attributes for each indicator. A detailed description of
 #'   all the attributes returned is provided at
 #'   \url{https://api.dhsprogram.com/rest/dhs/indicators/fields}
@@ -219,7 +220,8 @@ dhs_data <- function(countryIds=NULL,
 #' # A complete list of examples for how each argument to the indicator API
 #' # endpoint can be provided is given below, which is a copy of each of
 #' # the examples listed in the API at:
-#' \url{https://api.dhsprogram.com/#/api-indicators.cfm}
+#'
+#' https://api.dhsprogram.com/#/api-indicators.cfm
 #'
 #' \dontrun{
 #' dat <- dhs_indicators(countryIds="EG",all_results=FALSE)
@@ -285,7 +287,7 @@ dhs_indicators <- function(countryIds=NULL,
 #' @param page Allows specifying a page number to obtain for the API request. By
 #'   default the API will return page 1.
 #'
-#' @return Returns a `data.table` of 3 (or less if `returnFields` is provided)
+#' @return Returns a \code{data.table} of 3 (or less if \code{returnFields} is provided)
 #'   interfaces that have been added/updated or removed. A detailed description
 #'   of all the attributes returned is provided at
 #'   \url{https://api.dhsprogram.com/rest/dhs/uiupdates/fields}
@@ -302,7 +304,8 @@ dhs_indicators <- function(countryIds=NULL,
 #' # A complete list of examples for how each argument to the ui updates API
 #' # endpoint can be provided is given below, which is a copy of each of
 #' # the examples listed in the API at:
-#' \url{https://api.dhsprogram.com/#/api-uiupdates.cfm}
+#'
+#' https://api.dhsprogram.com/#/api-uiupdates.cfm
 #'
 #' \dontrun{
 #' dat <- dhs_ui_updates(lastUpdate="20150901",all_results=FALSE)
@@ -355,7 +358,7 @@ dhs_ui_updates <- function(lastUpdate=NULL,
 #' @param page Allows specifying a page number to obtain for the API request. By
 #'   default the API will return page 1.
 #'
-#' @return Returns a `data.table` of 2 (or less if `returnFields` is provided)
+#' @return Returns a \code{data.table} of 2 (or less if \code{returnFields} is provided)
[…]
