### 2018-09-24 8b12935
pkgdown corrections again
