+++ man/dhs_data.Rd
@@ -130,7 +130,7 @@ surveyYearStart="2006"
 # endpoint can be provided is given below, which is a copy of each of
 # the examples listed in the API at:
 
-https://api.dhsprogram.com/#/api-data.cfm
+# https://api.dhsprogram.com/#/api-data.cfm
 
 \dontrun{
 dat <- dhs_data(countryIds="EG",all_results=FALSE)
