+++ man/dhs_countries.Rd
@@ -94,7 +94,7 @@ dat <- dhs_countries(surveyType="MIS")
 # endpoint can be provided is given below, which is a copy of each of
 # the examples listed in the API at:
 
-https://api.dhsprogram.com/#/api-countries.cfm
+# https://api.dhsprogram.com/#/api-countries.cfm
 
 \dontrun{
 dat <- dhs_countries(countryIds="EG",all_results=FALSE)
