+++ man/dhs_info.Rd
@@ -58,7 +58,7 @@ dat <- dhs_info(infoType="version")
 # endpoint can be provided is given below, which is a copy of each of
 # the examples listed in the API at:
 
-https://api.dhsprogram.com/#/api-info.cfm
+# https://api.dhsprogram.com/#/api-info.cfm
 
 \dontrun{
 dat <- dhs_info(infoType="version",all_results=FALSE)
