### 2018-09-24 c04359b
corrections to geojson vignette
