+++ R/ui.R
@@ -53,7 +53,7 @@ rdhs_package_client_exists <- function() {
 #' # first create a config
 #' set_rdhs_config(
 #' config_path = "rdhs.json", global = FALSE, prompt = FALSE
-#'
+#' )
 #' # you can look at what you have set these to using \code{get_rdhs_config}
 #' get_rdhs_config()
 #'
