+++ man/get_rdhs_config.Rd
@@ -22,7 +22,7 @@ Returns the config being used by rdhs at the moment. This will
 # first create a config
 set_rdhs_config(
 config_path = "rdhs.json", global = FALSE, prompt = FALSE
-
+)
 # you can look at what you have set these to using \\code{get_rdhs_config}
 get_rdhs_config()
 
