+++ man/set_rdhs_config.Rd
@@ -111,6 +111,7 @@ config_path = "rdhs.json", global = FALSE)
 # otherwise we can do this by specifying prompt to FALSE
 set_rdhs_config(
 config_path = "rdhs.json", global = FALSE, prompt = FALSE
+)
 
 # you can look at what you have set these to using \\code{get_rdhs_config}
 get_rdhs_config()
