+++ R/config.R
@@ -136,7 +136,7 @@ handle_config <- function(config_path) {
 #' # the password. (not run as it requires a prompt)
 #' set_rdhs_config(email = "blah@gmail.com", project = "Blahs",
 #' config_path = "rdhs.json", global = FALSE)
-#' }
+#'
 #'
 #' # otherwise we can do this by specifying prompt to FALSE
 #' set_rdhs_config(
@@ -144,10 +144,8 @@ handle_config <- function(config_path) {
 #' )
 #'
 #' # you can look at what you have set these to using \code{get_rdhs_config}
-#' get_rdhs_config()
-#'
-#' # and you can view it prettily using \code{print_rdhs_config}
-#' print_rdhs_config(get_rdhs_config())
+#' config <- get_rdhs_config()
+#' }
 #'
 #'
 set_rdhs_config <- function(email = NULL,
