+++ man/set_rdhs_config.Rd
@@ -106,7 +106,7 @@ Setting up a configuration will enable API results to be cached, as
 # the password. (not run as it requires a prompt)
 set_rdhs_config(email = "blah@gmail.com", project = "Blahs",
 config_path = "rdhs.json", global = FALSE)
-}
+
 
 # otherwise we can do this by specifying prompt to FALSE
 set_rdhs_config(
@@ -114,10 +114,8 @@ config_path = "rdhs.json", global = FALSE, prompt = FALSE
 )
 
 # you can look at what you have set these to using \\code{get_rdhs_config}
-get_rdhs_config()
-
-# and you can view it prettily using \\code{print_rdhs_config}
-print_rdhs_config(get_rdhs_config())
+config <- get_rdhs_config()
+}
 
 
 }
