+++ DESCRIPTION
@@ -43,3 +43,4 @@ License: MIT + file LICENSE
 RoxygenNote: 6.1.0
 VignetteBuilder: knitr
 Language: en-GB
+Encoding: UTF-8
