### 2018-09-24 93be398
found it - added examples that get run first, so let's make the first example read in teh json if it's there
