### 2018-09-28 fb3107c
Corrections to examples relating to model dataset matching so we can run them without needing auth
