+++ R/api_endpoints.R
@@ -79,6 +79,7 @@
 #' @export
 #' @examples
 #'
+#' \dontrun{
 #' # A common use for the indicator data API will be to search for a specific
 #' # health indicator for a given country. For example to return the total
 #' # malaria prevalence according to RDT, given by the indicator ML_PMAL_C_RDT,
@@ -96,7 +97,7 @@
 #'
 #' # https://api.dhsprogram.com/#/api-data.cfm
 #'
-#' \dontrun{
+#'
 #' dat <- dhs_data(countryIds="EG",all_results=FALSE)
 #' dat <- dhs_data(indicatorIds="FE_FRTR_W_TFR",all_results=FALSE)
 #' dat <- dhs_data(surveyIds="SN2010DHS",all_results=FALSE)
@@ -206,7 +207,7 @@ dhs_data <- function(countryIds=NULL,
 #' @export
 #'
 #' @examples
-#'
+#' \dontrun{
 #' # A common use for the indicators data API will be to search for a list of
 #' # health indicators within a given characteristic category, such as anemia
 #' # testing, HIV prevalence, micronutrients etc. For example to return all the
@@ -223,7 +224,7 @@ dhs_data <- function(countryIds=NULL,
 #'
 #' # https://api.dhsprogram.com/#/api-indicators.cfm
 #'
-#' \dontrun{
+#'
 #' dat <- dhs_indicators(countryIds="EG",all_results=FALSE)
 #' dat <- dhs_indicators(indicatorIds="FE_FRTR_W_TFR",all_results=FALSE)
 #' dat <- dhs_indicators(surveyIds="SN2010DHS",all_results=FALSE)
@@ -295,6 +296,7 @@ dhs_indicators <- function(countryIds=NULL,
 #' @export
 #' @examples
 #'
+#' \dontrun{
 #' # The main use for the ui updates API will be to search for the last time
 #' # there was a change to the UI. For example to return all the
 #' # changes since 2018:
@@ -307,7 +309,7 @@ dhs_indicators <- function(countryIds=NULL,
 #'
 #' # https://api.dhsprogram.com/#/api-uiupdates.cfm
 #'
-#' \dontrun{
+#'
 #' dat <- dhs_ui_updates(lastUpdate="20150901",all_results=FALSE)
 #' dat <- dhs_ui_updates(f="html",all_results=FALSE)
 #' }
@@ -366,6 +368,7 @@ dhs_ui_updates <- function(lastUpdate=NULL,
 #' @export
 #' @examples
 #'
+#' \dontrun{
 #' # The main use for the info API  will be to confirm the version of the API
 #' # being used to providing the most current citation for the data.
 #'
@@ -377,7 +380,7 @@ dhs_ui_updates <- function(lastUpdate=NULL,
 #'
 #' # https://api.dhsprogram.com/#/api-info.cfm
 #'
-#' \dontrun{
+#'
 #' dat <- dhs_info(infoType="version",all_results=FALSE)
 #' dat <- dhs_info(infoType="citation",all_results=FALSE)
 #' dat <- dhs_info(f="html",all_results=FALSE)
@@ -456,6 +459,7 @@ dhs_info <- function(infoType=NULL,
 #' @export
 #' @examples
 #'
+#' \dontrun{
 #' # A common use for the countries API endpoint is to query which countries
 #' # ask questions about a given topic. For example to find all countries that
 #' # record data on malaria prevalence by RDT:
@@ -473,7 +477,7 @@ dhs_info <- function(infoType=NULL,
 #'
 #' # https://api.dhsprogram.com/#/api-countries.cfm
 #'
-#' \dontrun{
+#'
[…]
