### 2018-09-28 ae572ed
update_rdhs_config unexported function
