### 2018-09-28 #238 @OJWatson
Hi @LucyMcGowan,

Finally got there with the examples. So working through the comments.

1. Hope everything is going okay with :baby: and thank you again for finding the time for the review.

2. The test suite used to fail if you had not created a valid "rdhs.json" config file. I've touched the API tests won't fail by first creating a config if not found, as these tests can run without a DHS website account. This is included in commit [d66c8e](https://github.com/OJWatson/rdhs/commit/d66c8e55779a1e51f5d558d2f9963a248c5900e9)

 The other tests that require a login, however, will still require people to follow the testing vignette [here](https://ojwatson.github.io/rdhs/articles/testing.html)

3. There are now examples for all the exported functions, where they are needed. Where possible, these examples are run using the model datasets that the DHS website publishes, which anyone can download. This way you can use the main function in `ui.R`, e.g. `get_downloaded_datasets`

```
#' # get the model datasets included with the package
#' model_datasets <- model_datasets
#'
#' # download one of them
#' g <- get_datasets(dataset_filenames = model_datasets$FileName[1])
#'
#' # these will then be stored so that we know what datasets we have downloaded
#' d <- get_downloaded_datasets()
```

With regards to the examples for the  API functions, the long list of examples that you included for `dhs_data` was in there as it is a copy of all the examples that their API documentation provide, so I thought it would be nice to show how each one would be carried out using `rdhs`. I agree though this is not helpful, so I have added an initial couple of examples of common/useful use cases for each function with more descriptions as they go along. Then I explain what the long list of examples is so that people can link back to the DHS API website. For example for `dhs_ui_updates`

```
#' # The main use for the ui updates API will be to search for the last time
#' # there was a change to the UI. For example to return all the
#' # changes since 2018:
#'
#' dat <- dhs_ui_updates(lastUpdate="20180101")
#'
#' # A complete list of examples for how each argument to the ui updates API
#' # endpoint can be provided is given below, which is a copy of each of
#' # the examples listed in the API at:
#'
#' # https://api.dhsprogram.com/#/api-uiupdates.cfm
#'
#'
#' dat <- dhs_ui_updates(lastUpdate="20150901",all_results=FALSE)
#' dat <- dhs_ui_updates(f="html",all_results=FALSE)
#' }
```

I have, however, put these examples within `\dontrun{}` as their API can be temperamental.

4.  Thanks for linking `httr`'s way of doing this as this is much neater. Have implemented as well as adding it to the .Rbuildignore. This is included in commit [1ba973](https://github.com/OJWatson/rdhs/commit/1ba973ef34c9bfe553f49a0bea2cea38a25a6b05).

5. Completely agree. I have added an initial paragraph in both the README and the introduction vignette for this, which was included in commit [6c5554](https://github.com/OJWatson/rdhs/commit/6c5554983988b7cef033dab2c9c42e22b25cddf3)

6. Thank you picking these out and double thank you for making a PR with them.

I think that's everything. I've also made a few changes outside these comments, which clear up some of the issues that were outstanding and have added another example of how to use the "geojson" return type from API requests to build maps, shown [here](https://ojwatson.github.io/rdhs/articles/geojson.html)

---

Thank you again for the kind words about the package and for the comments and the suggestions. Definitely helped increase the usability of the package and thanks for linking the `add_line` function from `httr`, as i end up wanting that functionality a lot.

All the best,

OJ
