### 2018-09-30 #238 @annakrystalli
Hi @OJWatson, thanks for your swift response to @LucyMcGowan 's comments (and thanks @LucyMcGowan for them and your contributions, hope all is going well!). The package and documentation are really shaping up really nicely.

### Re (5)
- [ ] Because downloading data is such an important functionality, I feel introducing the user to the fact that they will have to register an rdhs account would be appropriate very early on,  in the README installation instructions, by including:
    >  To be able to download survey datasets from the DHS website, you will need to set up an account with them to enable you to request access to the datasets. Instructions on how to do this can be found [here](https://dhsprogram.com/data/Access-Instructions.cfm)

- [ ] Then, for more appropriate signposting to the testing vignette,  I think a summary of testing environment setup, with a link to the full testing vignette should be included in a **brief testing section of a `CONTRIBUTING.md`**. Here's a [template(https://github.com/ropensci/dotgithubfiles/blob/master/dotgithub/CONTRIBUTING.md) you can adapt. Sorry, I should have picked this up a bit earlier!. Depending on how likely you think users will need to know about the extra setup for testing, you could remove mention to it in the README, perhaps addinginstead a mention and link at the bottom to the `CONTRIBUTING.md` guidelinea. Or both.

Thoughts?
