### 2018-09-30 #238 @OJWatson
Hi @annakrystalli,

Thanks for the comments.

Completely agree. I have block quoted the registering for the DHS website bit within the README.md, so it now appears as the first section after the overview of the DHS, and is hopefully a lot clearer.

I have also moved mention of the testing vignette into the CONTRIBUTING.md, and have removed it from the README.md as i don't think it's that likely users will need to know about that step unless they are contributing. The CONTRIBUTING page has also been linked to within the pkgdown navbar to make it easier to find as well,

All the best and thanks for mentioning about the contributing.md and linking the dotgithubfiles repo as hadn't seen that before.

OJ
