### 2018-10-16 #238 @dosgillespie
Hi @OJWatson @annakrystalli

First I'm really sorry it has taken so long for me to look at this - I have been going through quite a busy period with work and family

Also I'm conscious that my review isn't very 'techie' as I've never developed a package myself and certainly don't have the know-how to comment on the details. But I am coming to this as someone interested in exploring the DHS data for the first time who is an experienced R user.

So my review is focused around my own worked example - in the R script below where I have annotated my thought process as I worked through and came up against snags.

To pull out two main comments:

- it would be good to have the info on how to register with the DHS and set your credentials in a prominent place right at the beginning of the instructions, so you can get that setup stuff out of the way before getting into exploring the data

- I kept coming up against the need to see some data documentation - to guide me to the relevant datasets and to explain variable names once I found them. My guess is that if I went through the DHS website I would find this - but it would be good to incorporate links to the relevant documentation or instructions on how to find it within the R package and its documentation.

Hope this helps
Duncan

`#First & Last Name:	Duncan Gillespie
#Project Title:	Development of health economic models for public health policies
#Email Address:	duncan.gillespie@sheffield.ac.uk

set_rdhs_config(email = "duncan.gillespie@sheffield.ac.uk",
                project = "Development of health economic models for public health policies",
                config_path = "rdhs.json",
                cache_path = "project_one",
                data_frame = "data.table::as.data.table",
                global = FALSE)

#set_rdhs_config(email = "rdhs.tester@gmail.com",
#                global = TRUE)

microbenchmark::microbenchmark(dhs_surveys(surveyYear = 1991),times = 1)

downloads <- get_datasets(datasets$FileName)
#Logging into DHS website...
#Error in names(filedatatypelist_DHS) <- paste0("filedatatypelist_", qdapRegex::rm_between(filedatatypelist_DHS_line,  :

# I needed to install digest as it was missing before I could get the install to work
#install.packages("digest")

# devtools::install_github("OJWatson/rdhs")
library(rdhs)

library(data.table)
library(ggplot2)

# I'm running this - trying to look for a way to tobacco and alcohol data
indicators <- dhs_indicators()
indicators[1,]

# try searching indicator definitions for 'tobacco'
tob_test <- indicators[grepl("tobacco", indicators$Definition)]
tob_test

# will try searching 'label' instead of 'definition'
tob_test <- indicators[grepl("tobacco", indicators$Label)]
tob_test

# try searching the tag id - 44
tob_test <- indicators[TagIds == 44]
tob_test

# looks like to calculate some crude rates of smoking I could somehow combine these results
  #Number of women (unweighted)
  #Number of men (unweighted)
  #Number of women who smoke cigarettes (unweighted)
  #Number of men who smoke cigarettes (unweighted)

# start with number of women who smoke
#AH_CIGA_W_UNW

# What are the tags
tags <- dhs_tags()

# Let's say we want to view the tags that relate to malaria
tags[grepl("Tobacco", tags$TagName), ]
#TagType TagName TagID TagOrder
#1:       0 Tobacco    44      650

# try not specifying a country id and instead specifying the indicator id above
#?dhs_data
data_w <- dhs_data(indicatorIds = "AH_CIGA_W_UNW")
data_m <- dhs_data(indicatorIds = "AH_CIGA_M_UNW")

# looks like I need the unweighted denominators from another file to calc the rates of smoking

data_den_w <- dhs_data(indicatorIds = "AH_TOBC_W_NUM")
data_den_m <- dhs_data(indicatorIds = "AH_TOBC_M_NUM")

# now to work out the matching variables for merging the datasets

#nrow(data)
#nrow(unique(data, by = c("CountryName", "SurveyYear")))

#nrow(data_den)
#nrow(unique(data_den, by = c("CountryName", "SurveyYear")))

# before merge only grab needed columns

data_w <- data_w[ , c("CountryName", "SurveyYear", "Value")]
setnames(data_w, "Value", "n_smokers")

data_m <- data_m[ , c("CountryName", "SurveyYear", "Value")]
setnames(data_m, "Value", "n_smokers")

data_den_w <- data_den_w[ , c("CountryName", "SurveyYear", "Value")]
setnames(data_den_w, "Value", "n_people")

data_den_m <- data_den_m[ , c("CountryName", "SurveyYear", "Value")]
setnames(data_den_m, "Value", "n_people")

# combine men and women data

data_w[ , sex := "women"]
data_m[ , sex := "men"]
data <- rbindlist(list(data_w, data_m), use.names = T)

data_den_w[ , sex := "women"]
data_den_m[ , sex := "men"]
data_den <- rbindlist(list(data_den_w, data_den_m), use.names = T)

# merge numberator and denominator
data <- merge(data, data_den, by = c("CountryName", "SurveyYear", "sex"), all.x = T, all.y = F, sort = F)

# now calculate smoking rates
data[ , rate := n_smokers / n_people]

# and try to visualise
ggplot(data) +
  geom_point(aes(x = reorder(CountryName, rate, max), y = rate, colour = SurveyYear, shape = sex)) +
  coord_flip() +
  xlab("country")

# briefly have a look for alcohol data

tags[grepl("Alcohol", tags$TagName), ]

indicators[grepl("alcohol", indicators$Description)]

# try this

## make a call with no arguments
sc <- dhs_survey_characteristics()
sc[grepl("Alcohol", sc$SurveyCharacteristicName), ]

# lets find all the surveys that fit our search criteria
survs <- dhs_surveys(surveyCharacteristicIds = 33)

# am looking for alcohol data and don't really want to trawl through all these

downloads <- get_datasets(datasets$FileName) # rather a lot of data being downloaded - dont want it - had to kill it

# just pick one survey
datasets <- dhs_datasets(surveyIds = "AL2008DHS", fileFormat = "flat")

downloads <- get_datasets(datasets$FileName)
#These requested datasets are not available from your DHS login credentials:
#  ALGE51FL.zip, ALBR50FL.ZIP, ALCR50FL.ZIP, ALHR50FL.ZIP, ALIR50FL.ZIP, ALKR50FL.ZIP, ALMR50FL.ZIP, ALPR50FL.ZIP, ALGC52FL.ZIP
#---

# I'm pretty sure I clicked on south africa - so lets try a survey from there
survs[DHS_CountryCode == "ZA"]

datasets <- dhs_datasets(surveyIds = "ZA1998DHS", fileFormat = "flat")
downloads <- get_datasets(datasets$FileName)

# having a look at the anemia worked example

surveychar <- dhs_survey_characteristics()
surveychar[grepl("alcohol", SurveyCharacteristicName, ignore.case=TRUE)]

surveys <- dhs_surveys(surveyCharacteristicIds = 33)
surveys[,.(SurveyId, CountryName, SurveyYear, NumberOfWomen, SurveyNum, FieldworkEnd)]

datasets <- dhs_datasets(surveyIds = surveys$SurveyId, fileType = "IR", fileFormat="flat")
datasets[, .(SurveyId, SurveyNum, FileDateLastModified, FileName)]

datasets$path <- unlist(get_datasets(datasets$FileName))
# ok so is downloading something now - I gave up prematurely...

#Downloading:
#  Malawi 2015 DHS Individual Recode Flat ASCII data (.dat) [MWIR7HFL.ZIP]
#Dataset download finished
#Downloading:
#  Tanzania 2003 AIS Individual Recode Flat ASCII data (.dat) [TZIR4AFL.zip]
#Dataset download finished
#Downloading:
#  Tanzania 2004 DHS Individual Recode Flat ASCII data (.dat) [TZIR4IFL.ZIP]
#Dataset download finished
#Downloading:
#  Tanzania 2007 AIS Individual Recode Flat ASCII data (.dat) [TZIR51FL.ZIP]
#Dataset download finished
#Downloading:
#  Tanzania 2010 DHS Individual Recode Flat ASCII data (.dat) [TZIR63FL.ZIP]
#Dataset download finished
#Downloading:
#  Tanzania 2015 DHS Individual Recode Flat ASCII data (.dat) [TZIR7AFL.ZIP]
#Dataset download finished

head(search_variable_labels(datasets$FileName[10], "alcohol")[,1:2])

# try this
cdpr <- readRDS("project_one/datasets/TZIR4AFL.rds")

dim(cdpr)

# rapid diagnostic test search
vars <- search_variable_labels(cdpr, search_terms = "alcohol")
# nope -
#Error in `[.data.frame`(filenames, , nm_type) :

# will finally try to use my experience to find something on this

tags[grepl("Mortality", tags$TagName), ]
#TagType               TagName TagID TagOrder
#1:       0       Adult Mortality    56      140
#2:       0 Child Mortality Rates    75      260
#3:       0    Maternal Mortality    38      570

mort_test <- indicators[TagIds == 56]

data_mort <- dhs_data(indicatorIds = "MM_AMRT_M_AMR")
# but what about age specific

data_mort <- dhs_data(indicatorIds = "MM_AMRT_M_DTH")
# I know that it must be possible to get these data for specific ages within countries but can't work out how to do it here
`
