### 2018-10-24 #238 @dosgillespie
Hi
I've tried redoing it in an rmd and writing that to a pdf so I can attach here - sorry if still a bit rubbish
[rdhs_review.pdf](https://github.com/ropensci/onboarding/files/2511466/rdhs_review.pdf)

I think the only specific extra installation step I had to do was to install the "digest" library, but I'm not sure if that was a problem with my system or not.

Duncan
