### 2018-10-29 #238 @OJWatson
Hi @dosgillespie and @annakrystall,

Thank you so much for the reviews and no worries about the timing - no rush from me and just glad to have the opportunity to have the code reviewed.

I've just come back from annual leave so haven't had time yet to go through the comments but i will try to respond sometime this week.

All the best,

OJ
