### 2018-11-09 7bc9591
#55 documentation error. Default = FALSE in clear cache
