### 2018-11-12 a0dc033
New updated Vignettes
