+++ vignettes/introduction.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "How to use rdhs?"
 author: "OJ Watson, Jeff Eaton"
-date: "2018-09-24"
+date: "2018-11-12"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{How to use rdhs?}
@@ -32,12 +32,17 @@ Install rdhs from github with `devtools`:
 library(rdhs)
 ```
 
+---
+
+ > Before starting the tutorial, if you wish to download survey datasets from the DHS website, you will need to set up an account with the DHS website, which will enable you to request access to the datasets. Instructions on how to do this can be found [here](https://dhsprogram.com/data/Access-Instructions.cfm). The email, password, and project name that were used to create the account will then need to be provided to `rdhs` when attempting to download datasets. You can still interact with the DHS API in section 1-2 below without having an account with the DHS website, however, you will need to create an account if you wish to go through steps 3-5. 
+
+---
 
 ## 1. Access standard indicator data via the API
 
-The DHS API has many *endpoints* that can be accessed using anyone of `dhs_<endpoint>()` functions. All exported functions within `rdhs` that start *dhs_* interact with a different with a different endpoint of the [DHS API](https://api.dhsprogram.com/). Their website gives great information about the different search terms and filters that can be used, and we have tried to include all of this within the documentation of each function.
+The DHS programme has published an API that gives access to a number of different data sets. These data sets include the standard health indicators that are available within [DHS STATcompiler](https://www.statcompiler.com/) as well as a series of meta data sets that describe the types of surveys that have been conducted as well as which raw dataset files are available from which surveys. Each of these data sets are described within the [DHS API website](https://api.dhsprogram.com/), and there are currently 12 different data sets available from the API. Each of these data sets can be accessed using anyone of `dhs_<>()` functions.  All exported functions within `rdhs` that start *dhs_* interact with a different with a different data set of the [DHS API](https://api.dhsprogram.com/). Their website gives great information about the different search terms and filters that can be used, and we have tried to include all of this within the documentation of each function.
 
-One of those endpoints, `dhs_data()`, interacts with the the published set of standard health indicator data calculated by the DHS. This endpoint allows use to retrieve a set of health indicators that have been sample weighted to give country, subnational estimates that can be further refined by education and wealth brackets. To do this we use the `dhs_data()` endpoint, which we can then either search for specific indicators, or by querying for indicators that have been tagged within specific areas.
+One of those functions, `dhs_data()`, interacts with the the published set of standard health indicator data calculated by the DHS. This data set contains a set of health indicators that have been sample weighted to give country, subnational estimates that can be further refined by education and wealth brackets. To do this we use the `dhs_data()` function, which we can then either search for specific indicators, or by querying for indicators that have been tagged within specific areas.
 
 
 ```r
@@ -47,25 +52,25 @@ indicators[1,]
 ```
 
 ```
-##                                                                                                           Definition
[…]
