### 2018-11-12 c52ecf9
pkgdown updates and new version number to reflect edits for ropensci
