+++ README.md
@@ -13,11 +13,11 @@ The Demographic and Health Surveys (DHS) Program has collected population survey
 
 ---
 
- > To be able to download survey datasets from the DHS website, you will need to set up an account with them to enable you to request access to the datasets. Instructions on how to do this can be found [here](https://dhsprogram.com/data/Access-Instructions.cfm). The email, password, and project name that were used to create the account will then need to be provided to `rdhs` when attempting to download datasets. 
+ > To be able to download survey datasets from the DHS website, you will need to set up an account with the DHS website, which will enable you to request access to the datasets. Instructions on how to do this can be found [here](https://dhsprogram.com/data/Access-Instructions.cfm). The email, password, and project name that were used to create the account will then need to be provided to `rdhs` when attempting to download datasets. 
 
 ---
 
-`rdhs` is a package for management and analysis of [Demographic and Health Survey (DHS)](www.dhsprogram.com) data. This includes functionality to:
+`rdhs` is a package for management and analysis of [Demographic and Health Survey (DHS)](https://www.dhsprogram.com) data. This includes functionality to:
 
 1. Access standard indicator data (i.e. [DHS STATcompiler](https://www.statcompiler.com/)) in R via the [DHS API](https://api.dhsprogram.com/).
 1. Identify surveys and datasets relevant to a particular analysis.
@@ -38,11 +38,11 @@ library(rdhs)
 
 ## Getting started
 
+* Request dataset access from the DHS website [here](https://dhsprogram.com/data/Access-Instructions.cfm).
+
 * Full functionality is described in the tutorial [here](https://ojwatson.github.io/rdhs/articles/introduction.html).
 
 * An example workflow using `rdhs` to calculate trends in anemia prevalence is available [here](https://ojwatson.github.io/rdhs/articles/anemia.html).
-
-* To run the full test suite (including tests for downloading datasets) see tutorial [here](https://ojwatson.github.io/rdhs/articles/testing.html).
 
 ## Basic Functionality
 
@@ -53,6 +53,8 @@ Obtain survey estimates for Malaria prevalence among children from the Democrati
 
 ```r
 dhs_indicators(indicatorIds = "ML_PMAL_C_RDT", returnFields=c("IndicatorId", "ShortName"))
+#> Warning in strptime(x, fmt, tz = "GMT"): unable to identify current timezone 'C':
+#> please set environment variable 'TZ'
 #>                             ShortName   IndicatorId
 #> 1 Malaria prevalence according to RDT ML_PMAL_C_RDT
 
@@ -61,9 +63,11 @@ dhs_data(countryIds = c("CD","TZ"), indicatorIds = "ML_PMAL_C_RDT", surveyYearSt
 #>                             Indicator  SurveyId SurveyYearLabel Value
 #> 1 Malaria prevalence according to RDT CD2013DHS         2013-14  30.8
 #> 2 Malaria prevalence according to RDT TZ2015DHS         2015-16  14.4
+#> 3 Malaria prevalence according to RDT TZ2017MIS            2017   7.3
 #>                 CountryName
 #> 1 Congo Democratic Republic
 #> 2                  Tanzania
+#> 3                  Tanzania
 ```
 
[…]
