### 2018-11-12 b3ec7fd
vignette typos and correction to country codes
