+++ vignettes/introduction.Rmd
@@ -40,7 +40,7 @@ library(rdhs)
 
 ## 1. Access standard indicator data via the API
 
-The DHS programme has published an API that gives access to a number of different data sets. These data sets include the standard health indicators that are available within [DHS STATcompiler](https://www.statcompiler.com/) as well as a series of meta data sets that describe the types of surveys that have been conducted as well as which raw dataset files are available from which surveys. Each of these data sets are described within the [DHS API website](https://api.dhsprogram.com/), and there are currently 12 different data sets available from the API. Each of these data sets can be accessed using anyone of `dhs_<>()` functions.  All exported functions within `rdhs` that start *dhs_* interact with a different with a different data set of the [DHS API](https://api.dhsprogram.com/). Their website gives great information about the different search terms and filters that can be used, and we have tried to include all of this within the documentation of each function.
+The DHS programme has published an API that gives access to a number of different data sets, which each represent one of the DHS API endpoints (e.g.  https://api.dhsprogram.com/rest/tags, or https://api.dhsprogram.com/rest/dhs/surveys). These data sets include the standard health indicators that are available within [DHS STATcompiler](https://www.statcompiler.com/) as well as a series of meta data sets that describe the types of surveys that have been conducted as well as which raw dataset files are available from which surveys. Each of these data sets are described within the [DHS API website](https://api.dhsprogram.com/), and there are currently 12 different data sets available from the API. Each of these data sets can be accessed using anyone of `dhs_<>()` functions.  All exported functions within `rdhs` that start *dhs_* interact with a different data set of the [DHS API](https://api.dhsprogram.com/). Their website gives great information about the different search terms and filters that can be used, and we have tried to include all of this within the documentation of each function. Each of these data sets
 
 One of those functions, `dhs_data()`, interacts with the the published set of standard health indicator data calculated by the DHS. This data set contains a set of health indicators that have been sample weighted to give country, subnational estimates that can be further refined by education and wealth brackets. To do this we use the `dhs_data()` function, which we can then either search for specific indicators, or by querying for indicators that have been tagged within specific areas.
 
@@ -52,22 +52,22 @@ indicators[1,]
 ```
 
 ```
-##                                                                                                            Definition
-## 1: Age-specific fertility rate for the three years preceding the survey for age group 10-14 expressed per 1,000 women
-##    NumberScale IndicatorType MeasurementType IsQuickStat  ShortName
-## 1:           0             I            Rate           0 ASFR 10-14
-##      IndicatorId    Level1 IndicatorTotalId          Level2 Level3
-## 1: FE_FRTR_W_A10 Fertility                  Fertility rates  Women
-##         SDRID IndicatorOldId TagIds DenominatorWeightedId
-## 1: FEFRTRWA10                                            
-##                                 Label IndicatorOrder
-## 1: Age specific fertility rate: 10-14       11763005
[…]
