### 2018-11-12 7afec29
typos and the rbuildinore correction
