### 2018-11-12 #238 @OJWatson
Hi @dosgillespie  and @annakrystalli ,

Apologies for the delay - was at a conference when i last messaged and had hoped to find some time during it to get round to this.

Firstly, thank you Duncan for taking time to go through it, and also for documenting your thoughts as you went through the introductory vignette. I found these really helpful as there were lots of little points that you made that helped me correct sections in the documentation and vignettes that aren't explained very well.

I have structured my response by responding to the comments I found while reading the review you wrote. This should answer the 2 main comments you had as well.

1. The note you wrote about having to separately install `digest` is interesting. I would have thought that would have been installed as part of installing `devtools`, so I can only think that maybe `digest` was deleted somehow between installing `devtools` and installing `rdhs`. Any other thoughts @annakrystalli

2. Thank you for highlighting that I have not defined what endpoints are when talking about the API. I don't think referring to them as endpoints in the introductory document is that helpful, so i have adjusted this text to mention more that the endpoints refer to the different datasets that you interact with from the API.

3. The link between the county ID's and the country names is not very clear as the DHS uses their own 2 letter code, which I don't think is standardised. It was in the second section of the introductory vignette, but this is not very obvious, so I've made another vignette just showing this, which is available at:
https://ojwatson.github.io/rdhs/articles/country_codes.html

4. You've mentioned wanting a data dictionary for the data sets. This is a good point and something that is sort of in the package, but again not very clearly explained or highlighted. In your review you mentioned that the searching functions (`search_variables` and `search_variable_labels`) are useful for
helping find what you are looking for within a data set, and are how we think it's easiest for people to query for specific questions. However, we didn't  highlight how you can quickly look at what the survey variables mean. To help with this I've lengthened section 4 in the introductory section to show the different ways that you can see the data dictionary for a given dataset.

5. You mentioned that it would be good to have the instructions on setting up an account first. I have now added a similar block quoted section to the one on the GitHub Readme to the introductory vignette, and have explained in it which sections of the vignette you can do without the login details set up.

---

I think that addresses the comments I could find. All the best and thank you again for the review,

OJ
