### 2018-11-19 #238 @dosgillespie
Hi folks

Yes all fine by me

Best wishes,
Duncan

On Wed, 14 Nov 2018 at 13:49, Anna Krystalli <notifications@github.com>
wrote:

> Thanks for your response @OJWatson <https://github.com/OJWatson>, just
> gonna run a couple of final checks and get back to you on your query (1).
>
> @dosgillespie <https://github.com/dosgillespie> please have a look at OJs
> responses and let us know if you are satisfied with the responses or
> whether there any further comments before signing off. Many thanks again
> for your time.
>
