### 2018-11-19 #238 @annakrystalli
Thanks @dosgillespie for signing off.

Almost there @OJWatson,

#### checks

So I've just run the **checks** in the build panel and get this error being triggered by an example.

```
Running examples in ‘rdhs-Ex.R’ failed
The error most likely occurred in:

> base::assign(".ptime", proc.time(), pos = "CheckExEnv")
> ### Name: read_zipdata
> ### Title: Read filetype from a zipped folder based on the file ending
> ### Aliases: read_zipdata
>
> ### ** Examples
>
>
> # get the model datasets included in the package
> model_datasets <- model_datasets
>
> # download just the zip
> g <- get_datasets(
+ dataset_filenames = model_datasets$FileName[1],
+ download_option = "zip"
+ )
Downloading:
ModelDatasetCountry ModelDatasetSurveyYear DHS Births Recode Stata dataset (.dta) [zzbr62dt.zip]
Dataset download finished
>
> # and then read from the zip. This function is used internally by rdhs
> # when using `get_datasets` with `download_option = .rds` (default)
> r <- read_zipdata(
+ g[[1]], pattern = ".dta"
+ )
Error in `levels<-`(`*tmp*`, value = if (nl == nL) as.character(labels) else paste0(labels,  :
  factor level [19] is duplicated
Calls: read_zipdata -> readfn -> factor
Execution halted
* checking for unstated dependencies in ‘tests’ ... OK
* checking tests ...
```

Then the tests hang.

If I run `checks()` in the command line, the checks abort from a different error, and theh tests hang again:
```
Running examples in ‘rdhs-Ex.R’ failed
The error most likely occurred in:

> base::assign(".ptime", proc.time(), pos = "CheckExEnv")
> ### Name: data_and_labels
> ### Title: Create list of dataset and its variable names
> ### Aliases: data_and_labels
>
> ### ** Examples
>
> # get the model datasets included with the package
> model_datasets <- model_datasets
>
> # download one of them
> g <- get_datasets(dataset_filenames = model_datasets$FileName[1])
Your datasets and API calls will be cached here:
   -> /Users/Anna/Library/Caches/rdhs
Your datasets will be downloaded using the following config:

Error in rep("*", nchar(config$password)) : invalid 'times' argument
Calls: get_datasets -> check_for_client -> print_rdhs_config -> paste0
Execution halted
```

#### tests

If I run the tests through the RStudio **build tab**, I get this repeating ad infinitum and have to force quit RStudio:
```
rdhs would like to write to files outisde of your R temporary
directory. This is so that your datasets and API calls are cached
between R sessions. Do you confirm rdhs to write to files outside
your R temporary directry? (Enter 1 or 2)

1: Yes
2: No
```

If I just run `test()` on the command line, I get the following interactive option again which asks me for a password whatever I respond (ie in both Yes and No case). My impression was that any tests requiring authorisation would just be skipped if credentials were not detected?

#### re `digest`

I'm not sure about `digest`. You've stated it as an import and it didn't appear as an issues in the initial checks. I uninstalled `digest` locally and installed rdhs from source with `devtools::install(dependencies = T)` and it worked fine for me (although had to restart R if I had devtools loaded for the namespacing to update). Also `devtools::install_github("OJWatson/rdhs")` worked fine also so can't reproduce @dosgillespie's issue..
