### 2018-11-19 #238 @OJWatson
oh wow - so many errors. Have had a look at them and have a few questions @annakrystalli

1. Firstly i'm intrigued that the `read_zipdata` error threw in the R build panel. What package version of `foreign` do you have installed as maybe this fails with a newer/older version?
2. The second fail in the examples I think makes sense to me so should have a fix for that.
3. For the tests have you created an `rdhs.json` credentials file in the testing directory? As am trying to work out the likely place where the need for user prompt has appeared from.

I'll hopefully have time to fix these today (however their API is down/slow at the moment so i might be speaking prematurely).

Thanks again and sorry these have sprung up,

OJ
