### 2018-11-19 #238 @annakrystalli
Hey @OJWatson , not to worry.

1) The version of `foreign` I have is 0.8-69.
3) No but my understanding was that tests that required the `rdhs.json` would be skipped if it wasn't available.  ie I should be able to test the functions that don't require authorisation and the rest should just skip.
