### 2018-11-20 4915de2
when printing config, check first if there is a password before obscuring it. It is possible for someone to create a config without a password and so this should be checked.
Associated tests and tidies
