### 2018-11-20 #238 @OJWatson
Hi @annakrystalli,

I was able to replicate the tests causing prompts, and I think i've found the cause and have patched that [(4915de2)](https://github.com/OJWatson/rdhs/commit/4915de29817b8b0b1460241d3b6b67eece659e31).

I'm struggling to replicate the first error relating to the example in `read_zipdata` though with your version of `foreign`. I've just checked in a new clone of the repo `v0.8-69` and can't get it to fail so i'm a little stumped. Does `devtools:::run_example("man/read_zipdata.Rd")` also throw the error when run in isolation? Maybe it was something in another example that caused it to break.

Sorry for not knowing better why that error was thrown, but I think the other error thrown in the examples has been addressed and the tests should skip correctly where authentication is required.

Thank you again for the help,

OJ
