### 2018-11-21 #238 @annakrystalli
Hey OJ,

So I still get this error whichever way I run the example:
```
Error in `levels<-`(`*tmp*`, value = if (nl == nL) as.character(labels) else paste0(labels,  :
  factor level [19] is duplicated
Calls: read_zipdata -> readfn -> factor
Execution halted
```
Is that the error you can't reproduce?

It looks like something is being coerced to a factor and levels being set that are not unique (eg see this [stackoverflow Q](https://stackoverflow.com/questions/38931194/warning-when-defining-factor-duplicated-levels-in-factors-are-deprecated). Does that make sense to you?

There's also a NOTE being thrown from the checks:

```
checking R code for possible problems ... NOTE
tempdir_check: possible error in tempdir(TRUE): unused argument (TRUE)
```
The tests all pass now though so 🙌. However, they are throwing a warning for me:
```
OS reports request to set locale to "French_Belgium.1252" cannot be honored
```
Seems related to [this](https://stackoverflow.com/questions/8145886/change-time-locale-for-r)?
