### 2018-11-22 e60f4a8
tempdir specify the TRUE is for the current version of tempdir's check argument for clarity.
