### 2018-11-22 3dae96a
simplify tempdir
