+++ R/utils.R
@@ -103,12 +103,7 @@ file_dataset_format <- function(file_format) {
 #' @noRd
 tempdir_check <- function() {
 
-  if (getRversion() < '3.5.0') {
-    temp <- tempdir()
-  } else {
-    temp <- tempdir(TRUE)
-  }
-
+  temp <- tempdir()
   return(temp)
 }
 
