### 2018-11-22 #238 @OJWatson
have just noticed that travis checks have started failing and i think i might have a better idea where that first error comes from now.
