### 2018-11-22 34f9f03
fixes to patch forchanges to haven's label class
