### 2018-11-23 ab19652
more places where we need to check on haven version
