+++ R/read_dhs_flat.R
@@ -346,6 +346,7 @@ read_dhs_flat <- function(zfile, all_lower=TRUE, meta_source=NULL) {
   dat[dct$name] <- Map("attr<-", dat[dct$name], "label", dct$label)
   haslbl <- unlist(lapply(dct$labels, length)) > 0
 
+  # match on haven package version
   if (packageVersion("haven") > "1.1.2") {
   dat[dct$name[haslbl]] <- Map(haven::labelled, dat[dct$name[haslbl]],
                                dct$labels[haslbl],
