### 2018-11-23 a56369e
again checkng on haven version within rbind_labelled so that the label attribute is correctly saved
