### 2018-11-23 a9a9b84
series of small changes places where we grab the "label" attribute from datasets so that it is matched exactly, otherwise it will grabs "labels" in some cases. Also made new unexported function for checking the dataset label type dataset_label_type, which is used in a few places to test if the dataset was read in with haven or foreign
