### 2018-11-23 609cf73
correction to new dataset_label_type
