### 2018-11-23 5f7f091
more edits to new function to reflect the reformat option
