### 2018-11-23 162e17b
another typo to new function
