+++ R/read_datasets.R
@@ -346,7 +346,7 @@ dataset_label_type <- function(data, stop = TRUE) {
              packageVersion("haven") <= "1.1.2") {
     type <- "labelled"
   } else if (any(lapply(data,attributes) %>%
-                 vapply(names,character(1)) =="label")) {
+                 lapply(names) %>% unlist() =="label")) {
     type <- "labelled"
   } else if (stop) {
     stop("Dataset does not have a label.table attribute or any ",
