### 2018-11-23 42a233e
still more changes to get all the tests and functionality right
