+++ man/as_factor.labelled.Rd
@@ -1,0 +1,37 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/rbind_labelled.R
+\name{as_factor.labelled}
+\alias{as_factor.labelled}
+\title{Archived dataset capable as_factor}
+\usage{
+as_factor.labelled(x, levels = c("default", "labels", "values", "both"),
+  ordered = FALSE, ...)
+}
+\arguments{
+\item{x}{Object to coerce to a factor.}
+
+\item{levels}{How to create the levels of the generated factor:
+\itemize{
+\item "default": uses labels where available, otherwise the values.
+Labels are sorted by value.
+\item "both": like "default", but pastes together the level and value
+\item "label": use only the labels; unlabelled values become \code{NA}
+\item "values: use only the values
+}}
+
+\item{ordered}{If \code{TRUE} create an ordered (ordinal) factor, if
+\code{FALSE} (the default) create a regular (nominal) factor.}
+
+\item{...}{Other arguments passed down to method.}
+
+\item{...}{Arguments passed on to \code{haven::as_factor}
+\describe{
+  \item{x}{Object to coerce to a factor.}
+}}
+}
+\description{
+Changes in `haven` have meant that `labelled` class are now referred to as
+`haven_labelled` classes. If `haven::as_factor` is used on old datasets they
+will fail to find the suitable method. rdhs::as_factor.labelled will work on
+old archived datasets that have a `labelled` class.
+}
