### 2018-11-23 bc5dc78
model_datasets is now loaded correctly so that functions work with. Fixes #60
