### 2018-11-26 a5c638d
Correct (maybe) way to make a data set internal and exported, by using both use_data and use_data_internal, as we want model_datasets to be both available to the user for demo purposes, and internal for the functonality.
