### 2018-11-26 212cbbf
better way t do as_factor.labelled
