### 2018-11-26 #238 @annakrystalli
Hey OJ! Sorry for delay in reply, I was away and took the whole weekend off! 😎

So re: reading in `.dta` files.

I tried running your code but had to make some slight adjustments to get it to work. Here's what I ran and the results of playing around a little:
##### setup file paths and tmp dir/file

``` r
file <- "https://dhsprogram.com/customcf/legacy/data/sample_download_dataset.cfm?Filename=ZZBR62DT.ZIP&Tp=1&Ctry_Code=zz&survey_id=0&doctype=dhs"
dest_file <- tempfile()
unzip_dir <- tempdir()
```

##### download data and unzip into unzip\_dir

``` r
download.file(file, destfile = dest_file)
unzip(dest_file, exdir = unzip_dir)
```

##### identify dta file

``` r
dta_file <- list.files(unzip_dir, pattern = ".DTA", full.names = T)
```

#### `foreign` (tested on versions 0.8-69 & 0.8-71) default - ERRORS

``` r
foreign::read.dta(dta_file)
#> Error in `levels<-`(`*tmp*`, value = if (nl == nL) as.character(labels) else paste0(labels, : factor level [19] is duplicated
```

#### `foreign` without converting factors - WORKS but attributes lost

``` r
dta_foreign <- tibble::as.tibble(foreign::read.dta(dta_file, convert.factors = F))
dta_foreign
#> # A tibble: 23,666 x 971
#>    caseid      bidx v000   v001  v002  v003  v004   v005  v006  v007  v008
#>  * <chr>      <int> <chr> <int> <int> <int> <int>  <int> <int> <int> <int>
#>  1 "        …     1 ZZ6       1     1     2     1 1.06e6     6  2015  1386
#>  2 "        …     2 ZZ6       1     1     2     1 1.06e6     6  2015  1386
#>  3 "        …     3 ZZ6       1     1     2     1 1.06e6     6  2015  1386
#>  4 "        …     4 ZZ6       1     1     2     1 1.06e6     6  2015  1386
#>  5 "        …     5 ZZ6       1     1     2     1 1.06e6     6  2015  1386
#>  6 "        …     6 ZZ6       1     1     2     1 1.06e6     6  2015  1386
#>  7 "        …     1 ZZ6       1     3     2     1 1.06e6     6  2015  1386
#>  8 "        …     1 ZZ6       1     4     2     1 1.06e6     6  2015  1386
#>  9 "        …     2 ZZ6       1     4     2     1 1.06e6     6  2015  1386
#> 10 "        …     1 ZZ6       1     4     3     1 1.06e6     6  2015  1386
#> # ... with 23,656 more rows, and 960 more variables: v009 <int>,
#> #   v010 <int>, v011 <int>, v012 <int>, v013 <int>, v014 <int>,
#> #   v015 <int>, v016 <int>, v017 <int>, v018 <int>, v019 <int>,
#> #   v019a <int>, v020 <int>, v021 <int>, v022 <int>, v023 <int>,
#> #   v024 <int>, v025 <int>, v026 <int>, v027 <int>, v028 <int>,
#> #   v029 <int>, v030 <int>, v031 <int>, v032 <int>, v034 <int>,
#> #   v040 <int>, v042 <int>, v044 <int>, v101 <int>, v102 <int>,
#> #   v103 <int>, v104 <int>, v105 <int>, v106 <int>, v107 <int>,
#> #   v113 <int>, v115 <int>, v116 <int>, v119 <int>, v120 <int>,
#> #   v121 <int>, v122 <int>, v123 <int>, v124 <int>, v125 <int>,
#> #   v127 <int>, v128 <int>, v129 <int>, v130 <int>, v131 <int>,
#> #   v133 <int>, v134 <int>, v135 <int>, v136 <int>, v137 <int>,
#> #   v138 <int>, v139 <int>, v140 <int>, v141 <int>, v149 <int>,
#> #   v150 <int>, v151 <int>, v152 <int>, v153 <int>, awfactt <int>,
#> #   awfactu <int>, awfactr <int>, awfacte <int>, awfactw <int>,
#> #   v155 <int>, v156 <int>, v157 <int>, v158 <int>, v159 <int>,
#> #   v160 <int>, v161 <int>, v166 <int>, v167 <int>, v168 <int>,
#> #   v190 <int>, v191 <int>, ml101 <int>, v201 <int>, v202 <int>,
#> #   v203 <int>, v204 <int>, v205 <int>, v206 <int>, v207 <int>,
#> #   v208 <int>, v209 <int>, v210 <int>, v211 <int>, v212 <int>,
#> #   v213 <int>, v214 <int>, v215 <int>, v216 <int>, v217 <int>, …
str(dta_foreign[,1:5])
#> Classes 'tbl_df', 'tbl' and 'data.frame':    23666 obs. of  5 variables:
#>  $ caseid: chr  "        1  1  2" "        1  1  2" "        1  1  2" "        1  1  2" ...
#>  $ bidx  : int  1 2 3 4 5 6 1 1 2 1 ...
#>  $ v000  : chr  "ZZ6" "ZZ6" "ZZ6" "ZZ6" ...
#>  $ v001  : int  1 1 1 1 1 1 1 1 1 1 ...
#>  $ v002  : int  1 1 1 1 1 1 3 4 4 4 ...
```

#### `haven` 2.0.0 - WORKS

``` r
dta_haven <- haven::read_dta(dta_file)
dta_haven
#> # A tibble: 23,666 x 971
#>    caseid      bidx v000   v001  v002  v003  v004   v005  v006  v007  v008
#>    <chr>      <dbl> <chr> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#>  1 "        …    1. ZZ6      1.    1.    2.    1. 1.06e6    6. 2015. 1386.
#>  2 "        …    2. ZZ6      1.    1.    2.    1. 1.06e6    6. 2015. 1386.
#>  3 "        …    3. ZZ6      1.    1.    2.    1. 1.06e6    6. 2015. 1386.
#>  4 "        …    4. ZZ6      1.    1.    2.    1. 1.06e6    6. 2015. 1386.
#>  5 "        …    5. ZZ6      1.    1.    2.    1. 1.06e6    6. 2015. 1386.
#>  6 "        …    6. ZZ6      1.    1.    2.    1. 1.06e6    6. 2015. 1386.
#>  7 "        …    1. ZZ6      1.    3.    2.    1. 1.06e6    6. 2015. 1386.
#>  8 "        …    1. ZZ6      1.    4.    2.    1. 1.06e6    6. 2015. 1386.
#>  9 "        …    2. ZZ6      1.    4.    2.    1. 1.06e6    6. 2015. 1386.
#> 10 "        …    1. ZZ6      1.    4.    3.    1. 1.06e6    6. 2015. 1386.
#> # ... with 23,656 more rows, and 960 more variables: v009 <dbl>,
#> #   v010 <dbl>, v011 <dbl>, v012 <dbl>, v013 <dbl+lbl>, v014 <dbl+lbl>,
#> #   v015 <dbl+lbl>, v016 <dbl>, v017 <dbl>, v018 <dbl+lbl>,
#> #   v019 <dbl+lbl>, v019a <dbl+lbl>, v020 <dbl+lbl>, v021 <dbl>,
#> #   v022 <dbl>, v023 <dbl+lbl>, v024 <dbl+lbl>, v025 <dbl+lbl>,
#> #   v026 <dbl+lbl>, v027 <dbl>, v028 <dbl>, v029 <dbl>, v030 <dbl>,
#> #   v031 <dbl>, v032 <dbl>, v034 <dbl+lbl>, v040 <dbl>, v042 <dbl+lbl>,
#> #   v044 <dbl+lbl>, v101 <dbl+lbl>, v102 <dbl+lbl>, v103 <dbl+lbl>,
#> #   v104 <dbl+lbl>, v105 <dbl+lbl>, v106 <dbl+lbl>, v107 <dbl+lbl>,
#> #   v113 <dbl+lbl>, v115 <dbl+lbl>, v116 <dbl+lbl>, v119 <dbl+lbl>,
#> #   v120 <dbl+lbl>, v121 <dbl+lbl>, v122 <dbl+lbl>, v123 <dbl+lbl>,
#> #   v124 <dbl+lbl>, v125 <dbl+lbl>, v127 <dbl+lbl>, v128 <dbl+lbl>,
#> #   v129 <dbl+lbl>, v130 <dbl+lbl>, v131 <dbl+lbl>, v133 <dbl+lbl>,
#> #   v134 <dbl+lbl>, v135 <dbl+lbl>, v136 <dbl>, v137 <dbl>, v138 <dbl>,
#> #   v139 <dbl+lbl>, v140 <dbl+lbl>, v141 <dbl+lbl>, v149 <dbl+lbl>,
#> #   v150 <dbl+lbl>, v151 <dbl+lbl>, v152 <dbl+lbl>, v153 <dbl+lbl>,
#> #   awfactt <dbl>, awfactu <dbl>, awfactr <dbl>, awfacte <dbl>,
#> #   awfactw <dbl>, v155 <dbl+lbl>, v156 <dbl+lbl>, v157 <dbl+lbl>,
#> #   v158 <dbl+lbl>, v159 <dbl+lbl>, v160 <dbl+lbl>, v161 <dbl+lbl>,
#> #   v166 <dbl+lbl>, v167 <dbl+lbl>, v168 <dbl+lbl>, v190 <dbl+lbl>,
#> #   v191 <dbl>, ml101 <dbl+lbl>, v201 <dbl>, v202 <dbl>, v203 <dbl>,
#> #   v204 <dbl>, v205 <dbl>, v206 <dbl>, v207 <dbl>, v208 <dbl+lbl>,
#> #   v209 <dbl+lbl>, v210 <dbl>, v211 <dbl>, v212 <dbl>, v213 <dbl+lbl>,
#> #   v214 <dbl>, v215 <dbl+lbl>, v216 <dbl+lbl>, v217 <dbl+lbl>, …
str(dta_haven[, 1:5])
#> Classes 'tbl_df', 'tbl' and 'data.frame':    23666 obs. of  5 variables:
#>  $ caseid: atomic          1  1  2         1  1  2         1  1  2         1  1  2 ...
#>   ..- attr(*, "label")= chr "case identification"
#>   ..- attr(*, "format.stata")= chr "%15s"
#>  $ bidx  : atomic  1 2 3 4 5 6 1 1 2 1 ...
#>   ..- attr(*, "label")= chr "birth column number"
#>   ..- attr(*, "format.stata")= chr "%8.0g"
#>  $ v000  : atomic  ZZ6 ZZ6 ZZ6 ZZ6 ...
#>   ..- attr(*, "label")= chr "country code and phase"
#>   ..- attr(*, "format.stata")= chr "%3s"
#>  $ v001  : atomic  1 1 1 1 1 1 1 1 1 1 ...
#>   ..- attr(*, "label")= chr "cluster number"
#>   ..- attr(*, "format.stata")= chr "%8.0g"
#>  $ v002  : atomic  1 1 1 1 1 1 3 4 4 4 ...
#>   ..- attr(*, "label")= chr "household number"
#>   ..- attr(*, "format.stata")= chr "%8.0g"

sessioninfo::session_info()
#> ─ Session info ──────────────────────────────────────────────────────────
#>  setting  value
#>  version  R version 3.4.4 (2018-03-15)
#>  os       macOS  10.14.1
#>  system   x86_64, darwin15.6.0
#>  ui       X11
#>  language (EN)
#>  collate  en_GB.UTF-8
#>  ctype    en_GB.UTF-8
#>  tz       Europe/London
#>  date     2018-11-26
#>
#> ─ Packages ──────────────────────────────────────────────────────────────
#>  package     * version date       lib source
#>  assertthat    0.2.0   2017-04-11 [1] CRAN (R 3.4.0)
#>  backports     1.1.2   2017-12-13 [1] CRAN (R 3.4.3)
#>  cli           1.0.1   2018-09-25 [1] CRAN (R 3.4.4)
#>  crayon        1.3.4   2017-09-16 [1] CRAN (R 3.4.1)
#>  digest        0.6.18  2018-10-10 [1] CRAN (R 3.4.4)
#>  evaluate      0.11    2018-07-17 [1] CRAN (R 3.4.4)
#>  forcats       0.3.0   2018-02-19 [1] CRAN (R 3.4.3)
#>  foreign       0.8-71  2018-07-20 [1] CRAN (R 3.4.4)
#>  haven         2.0.0   2018-11-22 [1] CRAN (R 3.4.4)
#>  hms           0.4.2   2018-03-10 [1] CRAN (R 3.4.4)
#>  htmltools     0.3.6   2017-04-28 [1] CRAN (R 3.4.0)
#>  knitr         1.20    2018-02-20 [1] CRAN (R 3.4.3)
#>  magrittr      1.5     2014-11-22 [1] CRAN (R 3.4.0)
#>  pillar        1.2.1   2018-02-27 [1] CRAN (R 3.4.3)
#>  pkgconfig     2.0.2   2018-08-16 [1] CRAN (R 3.4.4)
#>  R6            2.3.0   2018-10-04 [1] CRAN (R 3.4.4)
#>  Rcpp          1.0.0   2018-11-07 [1] CRAN (R 3.4.4)
#>  readr         1.1.1   2017-05-16 [1] CRAN (R 3.4.0)
#>  rlang         0.3.0.1 2018-10-25 [1] CRAN (R 3.4.4)
#>  rmarkdown     1.10    2018-06-11 [1] CRAN (R 3.4.4)
#>  rprojroot     1.3-2   2018-01-03 [1] CRAN (R 3.4.3)
#>  sessioninfo   1.1.0   2018-09-25 [1] CRAN (R 3.4.4)
#>  stringi       1.2.4   2018-07-20 [1] CRAN (R 3.4.4)
#>  stringr       1.3.1   2018-05-10 [1] CRAN (R 3.4.4)
#>  tibble        1.4.2   2018-01-22 [1] CRAN (R 3.4.3)
#>  utf8          1.1.3   2018-01-03 [1] CRAN (R 3.4.3)
#>  withr         2.1.2   2018-03-15 [1] CRAN (R 3.4.4)
#>  yaml          2.2.0   2018-07-25 [1] CRAN (R 3.4.4)
#>
#> [1] /Library/Frameworks/R.framework/Versions/3.4/Resources/library
```

Created on 2018-11-26 by the [reprex package](http://reprex.tidyverse.org) (v0.2.0).

So the problem looks like it is indeed emanating from `foreign`'s attempt to coerce to factors.  I'm wondering whether in might be related to changes in `R version 3.5.0`, namely that `factor(x, levels, labels)` now allow duplicated `labels` (not duplicated `levels`). You can get TRAVIS to explictly test for backcompatibility, might be worth it at this stage, given the differing behaviour you have set for `tempfile` also. Have a look at this: https://docs.travis-ci.com/user/languages/r#r-versions. Would be interested to know whether you manage to reproduce the `foreign` error when testing on on r `oldrel`. Overall `haven` seems to be the better approach IMHO, atleast for this use case as (if we have correctly identified the source of the error), it offers backcompatibility.

***
Re: `tempdir_check`, hmmmm I see what you mean. I think the function works to provide back compatibility but just throws this note when checking on using a version < 3.5.0.

I'm not sure what to suggest atm. I appreciate [there have been debates regarding](http://r.789695.n4.nabble.com/tempdir-may-be-deleted-during-long-running-R-session-td4732832.html) this which it seems `check = TRUE` option in `3.5` was designed to address.  But using it means differing behaviour according to R version (ie `tmpdir` retrievable with `3.5` but not retrievable with `3.4`). While this might be acceptable (but would atleast need pointing out in the documentation somewhere), I feel I'd need a bit more understanding of how exactly the user's workflow got messed up and whether this is something `rdhs` should be solving or better practice from the user.  My feeling is that tmp files are precisely that, temporary and designed to be deleted. So could you provide a bit more detail on how the error arose? ie

- What was the user trying to do? ie how did the `tmp/` folder get deleted?
- Which `rdhs` function was affected/implicated?

***

Re locale, I see. I'm going to have to seek some advice on this myself so will get back to you!

Thanks for sticking with the process! We'll get there 👍
