### 2018-11-26 433b67e
old release check for R using the corrected patching branch
