+++ .travis.yml
@@ -1,4 +1,7 @@
-language: R
+language: r
+r:
+  - oldrel
+  - release
 cache: packages
 
 r_packages:
