### 2018-11-26 16de6b4
gdal oldrel fails, maybe wrong libgdal
