### 2018-11-26 5de7517
The previous set of commits relating to the new haven label format in v.2.0 of haven have necessitated a lot of backwards compatibility changes. As a result of this and the development of functionslike rbind_labelled and our use of the labelled format as a whole, we are moving to the use of haven vs foreign for default reading in of dta. This should address the test issue in r <3.5
