### 2018-11-26 260f4da
Update to v0.6.0 and update pkgdown
