+++ .travis.yml
@@ -1,5 +1,6 @@
 language: r
 r:
+  - release
   - oldrel
 
 cache: packages
