### 2018-11-26 075fe83
having to split the oldrel and release travis builds as the oldrel requires the unstable ubuntugis, but this throws peculiar build errors in the release version
