### 2018-11-27 9955fcf
update_rdhs_config is improved/corrected now and is now exported as a result. And has associated test
