+++ NAMESPACE
@@ -28,6 +28,7 @@ export(read_zipdata)
 export(search_variable_labels)
 export(search_variables)
 export(set_rdhs_config)
+export(update_rdhs_config)
 importFrom(R6,R6Class)
 importFrom(magrittr,"%>%")
 importFrom(rappdirs,user_cache_dir)
