### 2018-11-27 211b2de
missing parameter documentation, and new pkgdown
