### 2018-11-29 973dced
read_dhs_dataset now can handle geospatial covariates
