### 2018-11-29 #238 @annakrystalli
OK, back with you.

So I pulled your changes and ran the tests. Getting a new failure and a couple of warnings (edited down):

``` r
devtools::test("/Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs")
#> Loading rdhs
#> Loading required package: testthat
#> Testing rdhs
───────────────────────────────────────────────────────────────────────────────────
⠏ |  0       | Miscellaneous utils
⠋ |  1       | Miscellaneous utils
⠙ |  2       | Miscellaneous utils
⠹ |  3       | Miscellaneous utils
⠸ |  4       | Miscellaneous utils
⠼ |  5       | Miscellaneous utils
⠴ |  6       | Miscellaneous utils
⠦ |  7       | Miscellaneous utils
⠧ |  8       | Miscellaneous utils
⠇ |  8     1 | Miscellaneous utils
⠏ |  9     1 | Miscellaneous utils
⠋ |  9   1 1 | Miscellaneous utils
⠙ | 10   1 1 | Miscellaneous utils
⠹ | 10   1 2 | Miscellaneous utils
⠸ | 10   1 3 | Miscellaneous utils
⠼ | 10   2 3 | Miscellaneous utils
⠴ | 10 1 2 3 | Miscellaneous utils
✖ | 10 1 2 3 | Miscellaneous utils
```

#### 1. The **warning** re: locale again:
``` r
#> test_miscellaneous.R:64: warning: different locales
#> OS reports request to set locale to "French_Belgium.1252" cannot be honored
```
I had a little play around with this and Im not sure whether it's testing what you want it to? ie on my system it fails to set the locale during testing (indeed it doesn't change the locale even if I run `Sys.setlocale("LC_TIME","French_Belgium.1252")` in the console.  The test doesn't fail though because the date remains parsable by `mdy_hms(date)`. Could you elaborate a little more on what it is you are trying to test here?

#### 2. Warning re: object 'model_datasets'
``` r
#> test_miscellaneous.R:114: warning: model datasets onAttach
#> object 'model_datasets' not found
```

I should note however that I don't get these warnings when running `devtools::check()`. Will seek some advice on this.

#### 3.**Error** : looks like another one requiring credentials that needs to be skipped

``` r
#> test_miscellaneous.R:115: error: model datasets onAttach
#> Login credentials for the DHS website have not been found.Please uset set_dhs_config() and provide your DHS email and project.
#> 1: rdhs::get_datasets("MWAR7ASV.ZIP") at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/tests/testthat/test_miscellaneous.R:115
#> 2: client$get_datasets(dataset_filenames, download_option = download_option, reformat = reformat,
#>        all_lower = all_lower, output_dir_root = output_dir_root, clear_cache = clear_cache,
#>        ...) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/ui.R:128
#> 3: handle_config(private$config$config_path) at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/client.R:345
#> 4: stop("Login credentials for the DHS website have not been found.", "Please uset set_dhs_config() and provide your DHS email and project.") at /Users/Anna/Documents/workflows/rOpenSci/editorials/rdhs/R/config.R:35

#> ══ Results ════════════════════════════════════════════════════════════════════════
#> Duration: 46.8 s
#>
#> OK:       59
#> Failed:   1
#> Warnings: 2
#> Skipped:  29
```

Created on 2018-11-29 by the [reprex package](http://reprex.tidyverse.org) (v0.2.0).
