### 2018-11-29 3acb396
wrap the locale test in warning catch incase the OS can't set foreign LC_TIME.
also forgot a skip_auth for the last miscellaneous test
