### 2018-11-29 5f326c1
add read.csv to namespace imports to remove note
