+++ R/rdhs.R
@@ -11,6 +11,7 @@
 #'
 #' @importFrom stats setNames
 #' @importFrom utils tail type.convert packageVersion unzip str capture.output
+#' @importFrom utils read.csv
 #' @importFrom rappdirs user_cache_dir
 #' @importFrom R6 R6Class
 #' @importFrom storr storr_rds
