### 2018-11-29 f253f0f
and NAMESPACE to reflect read.csv import
