+++ NAMESPACE
@@ -36,6 +36,7 @@ importFrom(stats,setNames)
 importFrom(storr,storr_rds)
 importFrom(utils,capture.output)
 importFrom(utils,packageVersion)
+importFrom(utils,read.csv)
 importFrom(utils,str)
 importFrom(utils,tail)
 importFrom(utils,type.convert)
