### 2018-11-29 #238 @OJWatson
Hi again, and sorry for the new errors :(

### 1. re: locale

Some users' OS may have different locale environment variable `LC_TIME`, which for one user in this [issue](https://github.com/OJWatson/rdhs/issues/37) lead to them being unable to parse date strings that were returned from the DHS API. So we have a function `mdy_hms`, which will change their `LC_TIME` env. variable to be equal to `C` for the purposes of parsing dates from the API. The following reprex (run on a windows)(should kind of show the problem a bit clearer a bit more information about what the locale change is for:

```
library(rdhs)

# the API returns dates in the following string
date <- "July, 15 2016 19:17:14"

# these will be correctly parsed if LC_TIME recognises it
strptime(date, format = "%B, %d %Y %H:%M:%S")
#> [1] "2016-07-15 19:17:14 UTC"

# which in this session is
locale_lc_time <- Sys.getlocale("LC_TIME")
locale_lc_time
#> [1] "English_United Kingdom.1252"

# if however it does not, e.g French_Belgium.1252
Sys.setlocale("LC_TIME","French_Belgium.1252")
#> [1] "French_Belgium.1252"
Sys.getlocale("LC_TIME")
#> [1] "French_Belgium.1252"
strptime(date, format = "%B, %d %Y %H:%M:%S")
#> [1] NA

# our function mdy_hms will change your LC_TIME to C before using strptime
mdy_hms(date)
#> [1] "2016-07-15 19:17:14 UTC"

# but will not change LC_TIME
Sys.getlocale("LC_TIME")
#> [1] "French_Belgium.1252"
```

The above, however, I was only able to do on my windows machine, and on my linux machine I get the same `cannot be honoured` message as you do. As a result, I've wrapped the test in a `tryCatch` to catch for a warning related to this, and to skip that test if it does throw the warning, i.e. the OS can't handle it.

### 2 & 3:
These both refer to a test I added recently, which was to address an issue where I had a forgotten to make `model_datasets`, which is a core data set the package uses when downloading data sets, an internal data set as well as one that should be exposed to the end user for demo purposes etc. This meant that a lot of functions would not work without `rdhs` being attached. I've slightly tweaked the test to better test this, and put an authentication skip on it as it does require auth.

---

I've just pushed these changes now into the master branch as well, so hopefully :crossed_fingers: these tests should then be handled okay now (famous last words)

Thanks again for everything, OJ
