### 2018-11-29 #238 @annakrystalli
> famous last words

...indeed! 😜

So, I asked for some advice from the rest of the editors and they hit up against another error when running the tests for the first time:

```
── 1. Error: post api tidy (@test_api_endpoints.R#354)  ────────────────────────
object 'conf' not found
1: .handleSimpleError(function (e)
   {
       handled <<- TRUE
       test_error <<- e
       options(expressions = expressions_opt_new)
       on.exit(options(expressions = expressions_opt), add = TRUE)
       e$expectation_calls <- frame_calls(11, 2)
       test_error <<- e
       register_expectation(e)
       e$handled <- TRUE
       test_error <<- e
   }, "object 'conf' not found", quote(eval(code, test_env))) at testthat/test_api_endpoints.R:354
```
Somehow I don't get this error, not sure why because I haven't set up credentials and `get_rdhs_config()` returns `NULL` for me. In any case, looking at the test itself I think there might be an error:

```r
test_that("post api tidy", {

  if (file.exists("rdhs.json")) {
  conf <- read_rdhs_config_file("rdhs.json")
  } else {
    expect_true(TRUE)
  }

  if (is.null(conf$email)){
    expect_true(file.remove("rdhs.json"))
  }

})
```
Feels like
```
 if (is.null(conf$email)){
    expect_true(file.remove("rdhs.json"))
  }
```
shouldn't run if `file.exists("rdhs.json") == FALSE` right? At the minute, even if the file doesn't exist in the first place and hence there is no `conf` object to interogate for `conf$email`, it still tries to, hence the error `object 'conf' not found`. So I think this needs correcting.

Interestingly, the secomd time they ran the tests, this error didn't turn up! Does the initial run create and cache a `rdhs.json` file somewhere? Perhaps after this particular test is run?
