### 2018-11-29 9cd306b
tiny change in case the test has no config
