### 2018-11-29 d692555
place the last api endpoint test cleanup better so that it catches for when all teh API functions have been skipped
