### 2018-11-29 #238 @OJWatson
Wooop! Okay and yes the above makes sense.

The API tests in `test_api_endpoints` will create an `rdhs.json` file in the current directory if there isn't one. Though if there wasn't one it won't have anything stored for `email` `project` `password` etc. So this tidy up was to make sure that any `rdhs.json` file created by the API tests is deleted if it will not be useful for the following tests that require auth. So i do a check on `conf$email`.

However, the following should have been inside the first `if` statement like so:

```
  if (file.exists("rdhs.json")) {
    conf <- read_rdhs_config_file("rdhs.json")
    if (is.null(conf$email)){
      expect_true(file.remove("rdhs.json"))
    }

  } else {
    expect_true(TRUE)
  }
```
However, I am a little confused though how they would have gotten to that test without creating an  `rdhs.json`. The only thing I can guess is that the API was currently down when they ran the tests, so all the API tests above that test would have been skipped and thus they wouldn't have made the "rdhs.json" file. Either way the changes above should then catch for that now, and these have been pushed.
