### 2018-11-30 f3a2643
Move info regarding account setup to getting started, embolden key info, remove quote
