### 2018-11-30 #238 @annakrystalli
Hey OJ,

Just been having a final review and look around the package, I made a small PR with some minor edits to the README, have a look and let me know what you think.

There's only one last point I want to raise. In `client_dhs()`, the API key is baked into the function as a default to argument `api_key`. While I appreciate it's a public key and see no reason for it not to be included in the source code, I feel a bit uncomfortable about it being so prominently user-facing. I think at some point I had even seen it turn up in errors when the function failed. I feel a better approach would be to set the default to NULL and then internally set it to the default one if no other API key is supplied to that argument.

Otherwise, I think we are good to 🚀 😃
