### 2018-11-30 efc0c9d
integrating new changes to README and pkgdown
