### 2018-12-02 9a2486e
Update README.md
md WIP removal
