### 2018-12-03 9813322
removal of api_key in text anbd moved to internal data object, and thus from documentation too
