+++ tests/testthat/test_ui.R
@@ -46,11 +46,10 @@ test_that("check on the use of the default client for API caching in extract", {
   dat <- dhs_countries(countryIds = "SN", all_results = FALSE)
 
   ## recreate the key
-
   key <- digest::digest(
     paste0(
       "https://api.dhsprogram.com/rest/dhs/countries?countryIds=SN",
-      "&f=json&apiKey=ICLSPH-527168",
+      "&f=json&apiKey=",api_key_internal,
     "all_results=", FALSE,
     collapse = ""
   ))
