+++ vignettes/client.Rmd
@@ -285,14 +285,14 @@ client
 ##     get_downloaded_datasets: function () 
 ##     get_root: function () 
 ##     get_variable_labels: function (dataset_filenames = NULL, dataset_paths = NULL, rm_na = FALSE) 
-##     initialize: function (config, api_key = "ICLSPH-527168", root = rappdirs_rdhs()) 
+##     initialize: function (config, api_key = NULL, root = rappdirs_rdhs()) 
 ##     save_client: function () 
 ##     set_cache_date: function (date) 
 ##     survey_questions: function (dataset_filenames, search_terms = NULL, essential_terms = NULL, 
 ##     survey_variables: function (dataset_filenames, variables, essential_variables = NULL, 
 ##   Private:
 ##     api_endpoints: data indicators countries surveys surveycharacteristics  ...
-##     api_key: ICLSPH-527168
+##     api_key: ********
 ##     cache_date: 2018-09-24 16:36:45
 ##     check_available_datasets: function (filenames) 
 ##     config: rdhs_config
