### 2018-12-03 #238 @OJWatson
Hi,

Completely agree - I have moved the actual string to be an internal package data object, and have removed all occurrences of it being written down explicitly in the code (although it is still in old commits).
As an aside the DHS programme said they wanted us to ship the package with that API key so that they could see how many people were using `rdhs` etc, but making it harder for people to get it makes sense.

Thanks again for everything,

OJ
