### 2018-12-04 e205700
update_rdhs_config did not correctly update data_frame args fix
