### 2018-12-04 b2ac300
update_rdhs_config error
