+++ vignettes/introduction.Rmd
@@ -28,7 +28,7 @@ Install rdhs from github with `devtools`:
 
 ```r
 # install.packages("devtools")
-# devtools::install_github("OJWatson/rdhs")
+# devtools::install_github("ropensci/rdhs")
 library(rdhs)
 ```
 
@@ -167,7 +167,7 @@ ggplot(resp[resp$CountryName %in% countries,],
   facet_wrap(~CountryName,ncol = 6) 
 ```
 
-<img src="https://github.com/OJWatson/rdhs/raw/development/vignettes/introduction_files/figure-html/unnamed-chunk-1-1.png" style="display: block; margin: auto;" />
+<img src="https://github.com/ropensci/rdhs/raw/development/vignettes/introduction_files/figure-html/unnamed-chunk-1-1.png" style="display: block; margin: auto;" />
 
 
 If we incorrectly entered a filter query (very possible), `rdhs` will let us know our request was invalid:
