+++ vignettes_src/client.Rmd
@@ -33,7 +33,7 @@ Install rdhs from github with `devtools`:
 
 ```{r gh-installation, warnings = FALSE, message = FALSE}
 # install.packages("devtools")
-# devtools::install_github("OJWatson/rdhs")
+# devtools::install_github("ropensci/rdhs")
 
 library(rdhs)
 ```
