### 2018-12-04 d9af1e7
last few link changes and removal of pkgdown reference pages for functions that don't exist anymore.
