### 2018-12-04 #238 @OJWatson
Hi again,

Have added the banner. Sorry for delay - had it sitting in the patching branch.

Also do you know how to get github to stop declaring the package's main code language as html? It keeps picking up the files in the `docs` folder, even though `.gitattributes` has specified `docs/* linguist-vendored`?

Many thanks again,

OJ
