### 2018-12-05 #238 @annakrystalli
Hi OJ, my guess (without digging into the history) is that the addition of the `.gitattributes` line came after the first commit of the `docs/`. Have a look at this [comment](https://medium.com/@monica_/hi-durim-kryeziu-thanks-for-reading-my-article-e1b47e781e04) and associated example [PR](https://github.com/durimkryeziu/jersey-2.x-webapp-for-servlet-container/pull/5) which describes the situation and how to correct it (I believe you just need to delete all docs, rebuild and recommit them for the linguist detector to ignore them). Let me know if that works!
