### 2018-12-05 d1d91a3
last bits for CRAN plus orcids
