### 2018-12-05 8b25737
package manifest with orcids
