### 2018-12-05 435cfb3
last cran edits, including changing email, one incorrect url in intro vignette and updating codemetar
