+++ DESCRIPTION
@@ -6,7 +6,7 @@ Authors@R:
     c(person(given = "OJ",
              family = "Watson",
              role = c("aut", "cre"),
-             email = "o.watson15@imperial.ac.uk",
+             email = "oj.watson@hotmail.co.uk",
              comment = c(ORCID = "0000-0003-2374-0741")),
       person(given = "Jeff",
              family = "Eaton",
@@ -23,7 +23,7 @@ Maintainer: OJ Watson <oj.watson@hotmail.co.uk>
 URL: https://ropensci.github.io/rdhs/
 BugReports: https://github.com/ropensci/rdhs/issues
 Description: Provides a client for (1) querying the DHS API for survey indicators
-  and metadata (https://api.dhsprogram.com/#/index.html), (2) identifying surveys
+  and metadata (<https://api.dhsprogram.com/#/index.html>), (2) identifying surveys
   and datasets for analysis, (3) downloading survey datasets from the DHS website,
   (4) loading datasets and associate metadata into R, and (5) extracting variables
   and combining datasets for pooled analysis.
