### 2018-12-05 #238 @stefaniebutland
> I'm definitely keen to write a blog post.

That's great to hear @OJWatson.
We could publish 2019-01-03 (Thursday 3rd) - the first post of the new year 🎉. That would mean submitting a draft for review by 2018-12-17 since I'm trying not to schedule any review work between Christmas and New Year.

This link will give you many examples of blog posts by authors of peer-reviewed packages so you can get an idea of the style and length you prefer: https://ropensci.org/tags/software-peer-review/.

Here are some technical and editorial guidelines for contributing a blog post: https://github.com/ropensci/roweb2#contributing-a-blog-post.

Happy to answer any questions.
