### 2018-12-10 3683c1b
These examples are too long for CRAN.
