+++ man/as_factor.labelled.Rd
@@ -34,7 +34,7 @@ old archived datasets that have a `labelled` class.
 For more details see \code{haven::as_factor}
 }
 \examples{
-
+\dontrun{
 # create a data.frame using the new haven_labelled class
 df1 <- data.frame(
 area = haven::labelled(c(1L, 2L, 3L), c("reg 1"=1,"reg 2"=2,"reg 3"=3)),
@@ -50,4 +50,5 @@ haven::as_factor(df1$area)
 
 # we can also use this on the data.frame by using the only_labelled argument
 haven::as_factor(df1, only_labelled = TRUE)
+}
 }
