### 2018-12-10 #238 @OJWatson
Hi,

Apologies for missing that!:

To-dos:

- [X]  Fix any links in badges for CI and coverage to point to the ropensci URL. We no longer transfer Appveyor projects to ropensci Appveyor account so after transfer of your repo to rOpenSci's "ropensci" GitHub organization the badge should be

## For JOSS submission
Once admin rights have been returned:
- [ ] Activate Zenodo watching for the repo
- [ ]  Submit to JOSS using the Zenodo DOI. We will tag it for expedited review.

^ I'm not planning on submitting to JOSS.

---

Sorry about missing that again - let me know if there is anything else,

OJ
