### 2018-12-10 2ec58ea
other tests are still too long for CRAN + codemetar
