### 2018-12-10 bab7ddd
more slow examples that CRAN doesn't like
