### 2018-12-10 1a91c98
another slow examples removed for CRAN
