+++ cran-comments.md
@@ -1,11 +1,10 @@
 ## Test environments
 * local Windows 7 x64, R 3.5.1
-* local Linux 4.15.0-39-generic (on travis-ci), R 3.5.1
+* local Linux 4.15.0-39-generic, R 3.5.1
 * local OS X, R 3.5.1
 
 ## R CMD check results
 There were no ERRORs or WARNINGs.
 
 ## Downstream dependencies
-
 There are no downstream dependencies.
