### 2018-12-10 0cccde0
all examples that require downloading data are now not run
