### 2018-12-10 #238 @annakrystalli
Great thanks for that! OK so last thing is to just add the rOpenSci badge to the README. (Sorry, this should have been done ages ago but I missed it)

Just ping when it's done and I'll close the issue 🙌😃
