### 2018-12-10 #238 @OJWatson
Sorry - that's my bad. I got confused between the footer and the badge. I've now added the badge to the readme.
