### 2018-12-19 #238 @stefaniebutland
Hi @OJWatson. I'm following up to see if you'd still like to contribute a blog post and to agree on a submission/publication date.

I'll be on vacation Dec 24 - Jan 1, so I could review a draft this week if submitted by Fri Dec 21, 9am Pacific, or review in the new year. If you are still interested, please date your post 2019-01-03 and we can modify that once it's ready.

This link will give you many examples of blog posts by authors of peer-reviewed packages so you can get an idea of the style and length you prefer: https://ropensci.org/tags/software-peer-review/.

Here are some technical and editorial guidelines for contributing a blog post: https://github.com/ropensci/roweb2#contributing-a-blog-post.

Happy to answer any questions.
