### 2019-01-04 #238 @stefaniebutland
Happy New Year @OJWatson. If you are still interested in writing a post about `rdhs`, I'm still interested in helping you get more eyes on your work 🙂. Please let me know at any time and we can establish a publication date that works for you.
