### 2019-01-04 #238 @OJWatson
Hi,

I am still keen to write a post about rdhs. Was hoping to do it over the holidays but had a paper bounce back just before so have been working on that. Am hoping to get to it in the following week though.

Thanks again,

OJ

From: Stefanie Butland
Sent: Friday 4 January, 21:40
Subject: Re: [ropensci/software-review] rdhs: submission for Ropensci (#238)
To: ropensci/software-review
Cc: Watson, Oliver, Mention

Happy New Year @OJWatson<https://github.com/OJWatson>. If you are still interested in writing a post about rdhs, I'm still interested in helping you get more eyes on your work 🙂. Please let me know at any time and we can establish a publication date that works for you.
